### 2024-09-25 #659 @heyairf
fireexposuR: Compute and Visualize Wildfire Exposure
Submitting Author Name: Air Forbes
Submitting Author Github Handle: @heyairf
Repository:  https://github.com/heyairf/fireexposuR
Version submitted: 1.0.0
Submission type: Standard

Version accepted: TBD

---

```
Package: fireexposuR
Title: Compute and Visualize Wildfire Exposure
Version: 1.0.0
Authors@R: c(person("Air", "Forbes", email = "amforbes@ualberta.ca",
    role = c("aut", "cre"), comment = c(ORCID = "0000-0002-9842-7648")),
    person("Jennifer", "Beverly", role = "aut"))
Description: This package computes wildfire exposure using methods from
    Beverly et al. (2010), Beverly et al. (2021), and Beverly and Forbes
    (2023).  It provides functions to standardize the mapping and
    visualization of wildfire exposure. This package requires
    pre-processing of data which is docomented in Forbes and Beverly
    (in preparation).
    geosphere,
    ggspatial,
    maptiles,
    MultiscaleDTM,
    terra,
    tidyselect,
    tidyterra
    knitr
```

This package was developed to share methodologies from existing research by modifying/expanding existing functions. The functions automate and standardize assessments to increase access and quality assurance in the calculation for interested users.

Users interested in wildfire risk assessments including but not limited to researchers, consultants, planners, land use decision makers.

There are no other R packages that fill this role.

n/a

[#652](https://github.com/ropensci/software-review/issues/652)

all pkgcheck's are currently passing
