### 2024-10-01 #659 @maurolepore
Hi @heyairf, thanks a lot for your submission. It's my pleasure to be the handling editor.

### Semantic tags for my comments

To help you track my comments I'll tag them with "ml" and a numbered sequence, e.g. ml01, ml02, and so on. Comments following bullets are for you to consider -- you may or may not respond to them. Comments following check-boxes are requests for some action -- please respond.

### Reviewers

- [ ] ml01. Can you please suggest three reviewers? Following our [guidelines](https://devguide.ropensci.org/softwarereview_editor.html#how-to-look-for-reviewers) I may appoint at most one; the other two will help me better understand the type of expertise you value.

### Editor checks

I'll post my checks in a few days. I already explored the package and looks in good shape. I appreciate your kindness in sharing your work with the rOpenSci community, and your willingness to learn a bit more about R package development. In that spirit I'll share  a number of suggestions to make the reviews as smooth as possible.
