### 2024-10-01 #659 @heyairf
Hi @maurolepore,
I am very much looking forward to the review process and opportunity to learn. Please note that the world of R package development is very new to me so my reviewer suggestions will reflect that. The following suggestions are people that participate in wildfire research projects, are proficient in R, and are, at varying levels, familiar with the methods that are in this package.

- Liz Chapman elizabeth.chapman@gov.bc.ca
- Jen Schmidt jischmidt@alaska.edu
- Brett Moore brett.moore@nrcan-rncan.gc.ca
