### 2024-10-10 #659 @maurolepore
Once again, thank for sharing your great work with the rOpenSci community.

Here are my checks and comments (based on the [editor's template](https://devdevguide.netlify.app/editortemplate)). Remember the items in bullets are just for your consideration, and the check-boxes require some action (sometimes just an explanation). The more items you can address before review the better. It will reduce the chance that reviewers will pick up in these relatively uninteresting structural issues and free them to focus on the more interesting aspects of the package.

Please feel free to ask for clarifications or help. If I don't hear from you before, I'll touch base in a couple of weeks.

---

##### Documentation

* ml01. The case for the package is made -- it provides an implementation in R of the methods described in three papers by Beverly et al. Consider giving a short description of each paper to help the users find the relevant one. For example, this clause leaves me wondering which paper explains how to prepare the data:

> "an accompanying paper details suggestions for data acquisition and preparation in accordance with various budget limitations and user experience levels."

- [ ] ml02. Please add a website for the package.

> We recommend creating a documentation website for your package using pkgdown.
> -- https://devdevguide.netlify.app/pkg_building.html#website

It sounds hard but all you need to do is to run `usethis::use_pkgdown_github_pages()` on the console:

> `use_pkgdown_github_pages()`: implements the GitHub setup needed to automatically publish your pkgdown site to GitHub pages:
> -- https://usethis.r-lib.org/reference/use_pkgdown.html

This allows "for an assessment of functionality and scope without installing the package", which should help the reviewers and users.

* ml03. The reference index (the list of all the exported functions and datasets of the package) can be made a little clearer by grouping functions. This package doesn't have many functions but it's still something to consider.

> When your package has many functions, use grouping in the reference, which you can do more or less automatically.
> https://devdevguide.netlify.app/pkg_building.html#function-grouping

More importantly, it would be nice to use references to "exposure" consistently in the names of the functions. Right now I see three ways to refer to exposure:

1. In full: *exposure*
2. As a suffix: adjust*exp*
3. In the middle: map*exp*class

Consider using a prefix rather than the alternatives. That way users can type the prefix and see all related functions thanks to auto-complete.

Also "We strongly recommend `snake_case` over all other styles unless you are porting over a package that is already in wide use". For example, `adjust_exp()` (or better even `exp_adjust()`) instead of `adjustexp()` (or `expadjust()`).

See Function and argument naming at https://devdevguide.netlify.app/pkg_building.html#function-and-argument-naming

- [ ] ml04. Can you please explain if you tried to include the example data in the package? I see some example data is created from scratch in README, in the vignette, and in many examples and tests (see also ml08). This is hard to maintain. If you later decide to use a different type of example data, doing that consistently would require changes in many places. I think somewhere you suggested the example data was too large to host in the package? This is something worth thinking about, since smaller test data correlates with more frequent testing and fewer bugs. Please explain your challenge and I'm more than happy to try help you find a solution.

##### Fit and overlap

- [ ] ml05. At https://github.com/ropensci/software-review/issues/659 I see: "This package was developed to share methodologies from existing research by modifying/expanding **existing functions**". Can you please explain where those existing functions live? I'd like to better understand the potential for overlap with some other package.

##### Tests

The tests could improve in a number of ways. I don't have first-had experience testing specifically spatial data so I'll try to find a reviewer who has (thanks for being proactive in https://discuss.ropensci.org/t/s4-data-for-testthat-and-examples/4007/5). For now I'll share some feedback that applies to tests in general.

- [ ] ml06. Please try to make each test run in under a second and let me know what challenges you face. Sometimes it's a matter of figuring out how to create a tiny toy dataset that allow you to test the properties you care about. For example, if you only care about the class of the output, the specific values of the output are irrelevant. This means that the test data can be unrealistically small AND useful, so you get useful and fast tests.

These two test-files are the slowest:

```
✔ |         20 | direxp [18.0s]
✔ |          7 | multidirexp [87.1s]
```

If you trully can't make tests run faster, then you may move them to a dedicated folder e.g. tests/testthat/slow and run those tests less frequently than every other test -- with `testthat::test_dir("tests/testthat/slow")` (or `testthat::test_dir(testthat::test_path("slow"))`).

References: https://testthat.r-lib.org/reference/test_dir.html and https://testthat.r-lib.org/reference/test_path.html

- [ ] ml07. In tests, try to find ways to run verbose functions in quiet mode -- unless the verbose output IS what you want to test (e.g. `expect_warning()`), but in those cases the expectations typically swallow the verbose output so that the test-report remains decluttered.

Here are the verbose functions and the output I see clutering the test-report:

```
adjustexp:
Proceed with caution: any adjustments to transmission distances should be further validated with observed fire history

direxp:
<SpatRaster> resampled to 500554 cells.
Value object provided has more than one feature, only the first point or polygon will be used.

mapexpcont:
<SpatRaster> resampled to 501264 cells.
```

See https://ropensci.org/blog/2024/02/06/verbosity-control-packages/

- [ ] ml08. (Overlaps with ml04 but may come with it's own challenges) Please see if there is a way to deduplicate the generation of test data. This will make it easier to maintain the package. If the tst data is useful to users, then see if you can provide it with the package or via a separate data-package. If it's not useful for users, then you may create test data in some other [Files relevant to testing](https://r-pkgs.org/testing-design.html#sec-tests-files-overview) e.g. (tests/testthat/setup.R). It can be hard to get your head around all these options. Have a look and let me know. I'm happy to help.

Specifially, I see this code-chunk multiple times:

```r
# test data ====================================================================
e <- c(45,55,495,505) * 10000
r <- terra::rast(resolution = 100, extent = terra::ext(e))
terra::values(r) <- sample(c(0,1), terra::ncell(r), replace = TRUE)
r <- terra::sieve(r, threshold = 50, directions = 4)
haz <- terra::sieve(r, threshold = 500, directions = 4)
```

- [ ] ml09. Please try to be specific about the type of errors you expect, e.g. `expect_error(f(), "calculation failed")` is better than `expect_error(f())`.

Here is one example from the package:

```r
  expect_error(direxp(2, pt))
```

> Note that the second argument to `expect_error()` is a regular expression — the goal is to find a short fragment of text that matches the error you expect and is unlikely to match errors that you don’t expect.
> -- https://mastering-shiny.org/scaling-testing.html?q=expect_error#key-expectations

* ml10. Suggestions based on the code chunk below:
	* It's best to be specific, and use `expect_message()`, `expect_warning()` rather than the catch-all `expect_condition().
	* It's best to use the `regexp` argument to ensure you get the specific contition you want to test.
	* The last empty call to `expect_condition()` seems like a mistake?

```r
test_that("mapexpclass() input checks and function messages work", {
  expect_condition(mapexplass(2, "loc", aoicrs)) # stopifnot() L67
  expect_condition(mapexpclass(expvals, "loc", aoicrs)) # stopifnot() L69
  expect_condition(mapexpclass(exp, "loc", 2)) # stopifnot() L71
  expect_condition(mapexpclass(exp, "loc", aoibig)) # stopifnot() L73
  expect_condition(mapexpclass(exp, "blah", aoicrs)) # matchargs() L75
  expect_condition()
})
```

* ml11. It's best for a function to always return an object of the same type.

>  If a function is type-stable ... you can predict the output type based only on the input types (**not their values**).
> -- https://design.tidyverse.org/out-type-stability.html

The code chunk below shows that `multidirexp()` is not type-stable. To predict whether the output is a "data.frame" or a "ggplot" we need to know the *value* of the argument `plot`. Consider creating separate functions for returning a "data.frame" and a "ggplot". For example:

- Option 1: `data |> multidirexp() |> plot_multidirexp()`
- Option 2: `data |> multidirexp() |> plot()` or `data |> multidirexp() |> ggplot2::autoplot()`. This requres understanding a little of the S3 object-oriented system (https://adv-r.hadley.nz/s3.html). Here are some examples using autoplot: https://forestgeo.github.io/fgeo.plot/reference/index.html

```r
# https://github.com/heyairf/fireexposuR/blob/main/tests/testthat/test-multidirexp.R
test_that("multidirexp() returns object with correct class", {
  expect_s3_class(multidirexp(exp, pts, plot = T), "ggplot")
  expect_s3_class(multidirexp(exp, pts), "data.frame")
})
```

* ml12. The comments are likely to fall out of sync with the code. It's best to make an effort to turn those comments into a formal testthat expectation.

```r

test_that("summexp() input checks work", {
  expect_condition(summexp(2)) #exp: right class
  expect_condition(summexp(exp, 2)) #aoi: right class
  expect_condition(summexp(exp, aoi, "blah")) # classify: match arg
})
```

##### Project management

- [ ] ml13. Please explain if you think it might be possible to reduce the size of the git repo. The entire directory is ~77M but the .git/ directory alone is ~76M. This means the current files are ~1M in total. This indicates previous versions of the repository did have large files that are no longer necessary. Do you really need those large files in the history? Removing files from the history of a repo is hard-core and irreversible but maybe there is a different, better place to store those large files so that the repo is light (and therefore faster to clone).

# The fireexposuR directory is quite large, so cloning is relatively slow
➜  git du -hd 1 fireexposuR
240K    fireexposuR/.Rproj.user
80K     fireexposuR/R
20K     fireexposuR/inst
76M     fireexposuR/.git
16K     fireexposuR/vignettes
24K     fireexposuR/.github
736K    fireexposuR/man
48K     fireexposuR/tests
77M     fireexposuR

# Almost all that "weight" comes from the .git/ directory alone
➜  git du -hd 1 fireexposuR/.git
4.0K    fireexposuR/.git/branches
75M     fireexposuR/.git/objects
32K     fireexposuR/.git/logs
8.0K    fireexposuR/.git/info
28K     fireexposuR/.git/refs
64K     fireexposuR/.git/hooks
76M     fireexposuR/.git
```

##### `@ropensci-bot` check

* ml14. The checks from the ropensci-bot show some noteworthy properties that may help find potential areas of improvement, such as removing unused dependencies, splitting or simplifying large or complex functions. In particular, the goodpractice results finds a call to the pattern `1:length(...)` that is best replaced with a call to `seq_len()`.

> Next, beware of iterating over 1:length(x), which will fail in unhelpful ways if x has length 0.
> -- https://adv-r.hadley.nz/control-flow.html?q=1:leng#common-pitfalls
