### 2024-10-15 #659 @heyairf
Hi @maurolepore,

Thank you so much for such detailed feedback! I will work on addressing these comments over the next few days and ask for further clarification if I need to. I very much appreciate all the resources you've included.
