### 2024-10-18 76e35b6
renamed functions, adjusted example and test data structure
