+++ _pkgdown.yml
@@ -1,0 +1,4 @@
+url: https://heyairf.github.io/fireexposuR/
+template:
+  bootstrap: 5
+
