+++ pkgdown.yml
@@ -1,0 +1,9 @@
+pandoc: 3.1.11
+pkgdown: 2.1.1
+pkgdown_sha: ~
+articles:
+  communityscale: communityscale.html
+last_built: 2024-10-18T21:52Z
+urls:
+  reference: https://heyairf.github.io/fireexposuR/reference
+  article: https://heyairf.github.io/fireexposuR/articles
