### 2024-10-21 #659 @heyairf
Hi @maurolepore,

I have attempted to addressed your feedback within my package! My replies to each of your comments are below. I expect that some may need to be addressed further, please let me know.

ml01
- The accompanying paper is currently in preparation and will be submitted after the R package review is complete so examples in the paper align with the r package after it has been reviewed. I have clarified this in the README and will be sure to update with the new citation when it is published.  I have also added a short description of the three papers that are linked, thank you for your suggestion!

ml02
- Thank you for the links. I have added a website for the package following your instructions!

ml03
- I have changed the function names so they use snake case and standardized all names to a prefixed “fire_exp”. Some functions have been split into more than one to address your feedback comment ml11. I have not officially grouped them, but now the function names organize them in a more meaningful order.

ml04/ml08
- This has absolutely been my biggest challenge in this process that I have struggled to find help with the most. Originally I tried to use real data that had very large file sizes but soon found that that would not be an option which is why I switched to generating the example data with code. It was most challenging to figure out because of the spatial data types and inability to save them as .Rdata files like almost every example on the internet. I have made some more changes to address this based on all your very helpful feedback! I am sure there is still lots of room for improvement, but I have now saved the main example data as a .tif file in inst/extdata instead of regenerating it on the fly in every single example and test document.

ml05
- My apologies, to clarify, the existing functions in question were made and used by only me. I have been using them in my own research and to save myself time when helping others run assessments. The edits and modifications I made were to make them more robust to share with a larger audience via the R package (i.e. the documentation and input checks).

ml06
- With your feedback I believe I have now reduced the testing time to under 4 seconds for each function while still maintaining test coverage. I will continue to think about ways to reduce the size/extent of the test data to speed these up further.

ml07
- I have added suppressMessages() to the few tests that output non-error messages to remove the clutter.

ml08
- see reply to ml04

ml09/ml10
- This feedback was so helpful for my overall understanding of how tests should be designed! I believe I have made significant improvements to address this in all of the test files.

ml11
- I have split some functions to address this, but perhaps it can be split further still. For example, `fire_exp_dir()` now returns a spatial feature, that can then be used in `fire_exp_dir_plot()` which has the option to return a plot or a map.

ml12
-   The comments have been removed now that the expected error messages (from feedback at ml09/ml10) are more meaningful

ml13
- You are absolutely correct that there was originally large files saved with the package for the examples in early stages of the project. I will remove them from the repo this week with some support from the research and training support at my university.

ml14
- I have updated this line to use seq_len()
