### 2024-10-23 7f5d801
Update fire_exp_map_class.R
