### 2024-10-24 cd7ef05
Update fire_exp_extract_vis.R
