### 2024-10-24 9f9745f
fix workshop issues
