### 2024-10-24 2146962
Update cbind to clear clutter in output
