+++ R/fire_exp_extract.R
@@ -49,7 +49,8 @@ fire_exp_extract <- function(exposure,
     ext1 <- terra::extract(exp, values, fun = mean, bind = TRUE) %>%
            dplyr::rename(mean_exp = exposure)
     ext2 <- terra::extract(exp, values, fun = max, bind = TRUE) %>%
-      dplyr::rename(max_exp = exposure)
+      dplyr::rename(max_exp = exposure) %>%
+      dplyr::select("max_exp")
     ext <- cbind(ext1, ext2)
   }
   return(ext)
