+++ R/fire_exp_extract.R
@@ -50,7 +50,7 @@ fire_exp_extract <- function(exposure,
            dplyr::rename(mean_exp = exposure)
     ext2 <- terra::extract(exp, values, fun = max, bind = TRUE) %>%
       dplyr::rename(max_exp = exposure) %>%
-      dplyr::select("max_exp")
+      dplyr::select("exposure")
     ext <- cbind(ext1, ext2)
   }
   return(ext)
