### 2024-10-24 dcf23cb
Adjust map colors
