### 2024-10-24 #659 @maurolepore
@heyairf thanks so much for your effort and your quick response. Great work.

I'll start reaching out to potential reviewers.

* ml13 Great you're taking initiative. For now I suggest to keep this repo as is. Before you push a re-written Git repo to GitHub I would like to first discuss with the editor board to gather alternative opinions. Until it seems prudent to practice in a local copy of the repo.

----

Here is a tiny new comment.

* ml15. In the tests I see this warning:

```r
Warning (test-fire_exp_extract_vis.R:43:3): fire_exp_extract_vis() runs when input conditions are met
Use of .data in tidyselect expressions was deprecated in tidyselect 1.2.0.
i Please use `"class"` instead of `.data$class`
```

The problem is in [R/fire_exp_extract_vis.R#L137](https://github.com/heyairf/fireexposuR/blob/01b6686c337c21e04eabef8f2b7d3980c2be9e88/R/fire_exp_extract_vis.R#L137). You can fix using use `"class"` instead of `.data$class`.

```r
      tidyr::drop_na("class")
```
