### 2024-10-25 #659 @maurolepore
ml13. I already discussed with the editorial board and we agree in that removing the large files would be beneficial, particularly before we engage the reviewers.

Here are some tools you may consider:

  * https://github.com/newren/git-filter-repo/
  * https://github.com/rtyley/bfg-repo-cleaner

Understanding the goal and knowing about a couple of tools should help you find examples and better support. If that's insufficient please let me know and I'm happy to help.

Remember to backup your original repo and play with copies of it until you're comfortable with the result. If
