### 2024-10-26 fbc8812
added rOpenSci badge
