### 2024-10-26 ff4c1dc
added input check
