### 2024-10-26 1afb8a4
updated code for changes in tidyselect
