+++ DESCRIPTION
@@ -4,7 +4,7 @@ Title: Compute and Visualize Wildfire Exposure
 Version: 1.0.1
 Authors@R: c(person("Air", "Forbes", email = "amforbes@ualberta.ca", 
     role = c("aut", "cre"), comment = c(ORCID = "0000-0002-9842-7648")),
-    person("Jennifer", "Beverly", role = "aut"))
+    person("Jennifer", "Beverly", role = "ths"))
 Description: This package computes wildfire exposure using methods from
     Beverly et al. (2010), Beverly et al. (2021), and Beverly and Forbes
     (2023).  It provides functions to standardize the mapping and
