### 2024-10-26 #659 @heyairf
ml13. thank you for your suggestions and investigation into where the problem is! I received the same links from the digital library resources at my university. I will try to arrange to meet with someone to go through the process of applying either method to my repo. This is something I'd rather have some in-person (or at least live) assistance with as I'm still learning... I expect I will be able to find someone to help me with on Monday! I will drop a note once the large files have been dealt with.

Regarding any potential Conflict of Interests from the recommended reviewers. I have not personally collaborated with any of the suggested reviewers on any previous projects. I will note, however, that Liz Chapman has worked on projects with my academic supervisor in the past. My supervisor is listed as an author on this project as their support and guidance to ensure the functions are properly representing the published methods they are automating was invaluable. I have updated the description and citation files to be explicit that I am the main author and contributor to the code and package development after reading this article: https://journal.r-project.org/articles/RJ-2012-009/.
