### 2024-10-31 #659 @heyairf
Hi @maurolepore ,

I got some help today using the bfg repo cleaner... I think we cleared the large tif files from the git history. would you be able to confirm that for me?

Thanks! Air
