### 2024-11-01 #659 @maurolepore
@heyairf great work! The repo is now about 10 times smaller. It was reduced from 80MB to 9M.

NOW

```
➜  git du -hd 1 fireexposuR
84K	fireexposuR/R
232K	fireexposuR/inst
7,9M	fireexposuR/.git
16K	fireexposuR/vignettes
28K	fireexposuR/.github
608K	fireexposuR/man
52K	fireexposuR/tests

9,0M	fireexposuR
```

BEFORE

```
➜  git du -hd 1 fireexposuR-backup
84K	fireexposuR-backup/R
232K	fireexposuR-backup/inst
79M	fireexposuR-backup/.git
16K	fireexposuR-backup/vignettes
28K	fireexposuR-backup/.github
608K	fireexposuR-backup/man
52K	fireexposuR-backup/tests

80M	fireexposuR-backup
```
