### 2024-11-12 #659 @heyairf
No problem at all, @maurolepore

In the meantime, is it okay if I make some small changes as issues come up? My colleagues have been using the package with their own data and I'd like to make some improvements based on their feedback.

Air
