### 2024-11-12 #659 @maurolepore
Dear @ronnyhdez, thanks again accepting so quickly to review this submission.

Happily the package already got some useful feedback and we're wondering whether to push it to `main` or a separate `branch`. What makes the difference is whether or not you already started you review.

Please do not feel pressed in any direction. I'm neither asking you to rush nor to slow down. I'm just checking the status to plan the best merge-strategy.

With that long context the question is this:

Have you already started your review (or are you planning to start it soon -- say before the end of this week)?
