### 2024-11-12 ef92f1b
added rOpenSci badge and updated text to match function name updates
