### 2024-11-13 #659 @ronnyhdez
Hi @heyairf and @maurolepore!

There is no problem with me if you merge the changes to the main branch.

I do have some notes about my review but mostly on documentation. I will continue this weekend with the review, so I could use the repo with new changes.

Thanks for letting me know!
