### 2024-11-13 #659 @maurolepore
@ronnyhdez thanks so much for the update and your flexibility.

@heyairf I hope the last few comments help you decide which strategy is best to ensure you can capitalize on the feedback you got while also ensuring Ronny's work remains relevant and smooth.

Let me know if anything is unclear or you need some other help.
