### 2024-11-18 #659 @ronnyhdez
Hi @heyairf and @maurolepore! Thanks for the chance to review this package! Here are my suggestions.

@heyairf if there is something that is not clear, please let me know. This is my first review, so I'm also learning about the process :smiley:

---

It would be helpful to clarify how coordinate reference systems are handled
throughout the package, particularly whether transformations affect only the
visualization outputs or also the underlying analyses. For example, the use of
tiles involves the re-projection to EPSG:3857, which is necessary for
visualization. Given that the target audience of this package seems to be
wildfire scientists, documenting these spatial operations would enhance the
functions' usability.

Also, it would be helpful to indicate the geographical scope of the implemented
methods. From the references included in the README, it seems that the studies
were all conducted in Alberta, Canada. If the methods are specifically calibrated
for this region's characteristics, this should be explicitly stated in the
package documentation to help users assess the applicability to their study areas.

While references to scientific publications provide valuable methodological
context, the function documentation would be more effective if it primarily
focused on usage and implementation details, with citations serving as
supplementary information. This would enhance the user experience by providing
immediate access to practical implementation guidance.

I would suggest adding a CITATION file (see rOpenSci guidelines
[here](https://devguide.ropensci.org/pkg_building.html#citation-file)) to help
users properly acknowledge your work when using the package in their research.

The checkpoiints above state that there should be a 'contribution guidelines'.
[Here](https://devguide.ropensci.org/maintenance_collaboration.html#contributing-guide)
is some information that can guide you about how to include it.

#### Specific thoughts

Here I have some comments/suggestions to specific sections of the package:

##### README

> "This package automates the methods previously documented in a series of scientific publications.

>    Beverly et al. 2010
        Introduces wildfire exposure and wildfire transmission distances for community scale assessments
    Beverly et al. 2021
        Validation of the wildfire exposure metric at a landscape scale with observed fire history
    Beverly and Forbes 2023
        Directional vulnerability assessment and methodology"

It's great to have references to published papers in a package. Nonetheless, the
documentation should focus on providing clear, step-by-step instructions for
using the package, rather than primarily redirecting users to scientific papers
for basic usage information.

I suggest including brief explanations of the methods directly in the README,
along with the citations. This would allow users to understand
the package's functionality immediately from the documentation, while still
having access to academic papers for a deeper understanding. Moreover, one of the
[papers](https://doi.org/10.1071/WF09071) is behind a paywall, so if a user
does not have access to the Journal, it will complicate the adoption of the
package (If the package is documented as papers first, before using the code).

> The functions in this package require the pre-preparation of data; an
accompanying paper (currently in preparation) details suggestions for data
acquisition and preparation in accordance with various budget limitations and
user experience levels."

This is also related to the previous comment. The scientific publication would
make your tools more robust, and for sure it should be included in the
package documentation. But try to prioritize practical usage guidance
explanations rather than redirecting users to a paper.

Think about a potential user of your package. They need to analyze wildfire
exposure and found your package. The first thing would be to glimpse the
documentation and give it a try just to check if it works and what kind of
outputs it can produce. A summary explanation with some example code will help
that user to perform a skim of the package and make the decision to use or not.

So, I wouldn't redirect them to a scientific paper as the first step. Make the
user explore your package and once they have tried, give them the possibility
to go for further details on the published papers.

Think of the README as an abstract: it should concisely present the package's
purpose and basic workflow. Users should be able to quickly assess if the
package meets their needs through example code and the workflow, before
diving into the underlying scientific publications for advanced theory
understanding.

##### Vignettes

The package already have vignettes which is great to document and explain
further how an user can benefit from using the package. There are no rules on
how documentation should be arrange in vignettes, but the general trend I see
and find useful is to have:

- **index page**: Almost the same as the README. The first point of entry of an
user could be the GitHub repository or the package website. So, to have the
"abstract" on both sides is good.

- **Get started page**: This would be the place to explain the workflow of the
package. I would try to not expand to much here about scientific details but
rather on the workflow and steps to get the primary functionality of the package
explained.  The deeper topics would be referenced to a vignette article.

- **Articles**: This is the section I would use to explain further details and
deeper concepts that were not covered on the "Get started" page. If an user
wants to go deeper in understanding specific details of methods, sciencie and
code, this is the section to look for that information.

Some examples of this arrangement are [dplyr](https://dplyr.tidyverse.org/articles/dplyr.html)
and [gt](https://gt.rstudio.com/index.html)

##### communityscale vignette

I liked this vignette because it already have the primary functionality and
workflow of the package. I would use it as the 'Get started' page. It takes the
user through a set of steps to check quickly how the package works.

About the code snippets, I suggest using complete names for the objects and
linespaces to separate steps. For example, the first chunk of code in the
`communityscale` vignette looks like this:

library(terra)
# read example hazard data ----------------------------------
filepath <- "extdata/hazard.tif"
haz <- terra::rast(system.file(filepath, package = "fireexposuR"))
# read example AOI
filepath <- "extdata/builtsimpleexamplegeom.csv"
g <- read.csv(system.file(filepath, package = "fireexposuR"))
aoi <- terra::vect(as.matrix(g), "polygons", crs = haz)
# generate random points
pts <- terra::spatSample(aoi, 200)
#' # ----------------------------------------------------------
```

I would change it the following way:

library(terra)

# Path to sample hazard data
hazard_file_path <- "extdata/hazard.tif"

# Read sample hazard data
hazard <- rast(system.file(file_path, package = "fireexposuR"))

# Read example Area of Interest (AOI)
aoi_file_path <- "extdata/builtsimpleexamplegeom.csv"
aoi_data <- read.csv(system.file(aoi_file_path, package = "fireexposuR"))
aoi_polygon <- vect(as.matrix(aoi_data), "polygons", crs = hazard)

# Generate random points
points <- spatSample(aoi_polygon, 200)
```

That way, it would be easier (less cognitive load) for a new user to
follow the examples and the objects being created and transformed within. I
suggest implementing this in all the user documentation.

If there is a need to explain further something on this 'Get started' page, I
would create an article, but I would leave in this 'Get Started' page, a link
to that more detailed explanation.

Further explanations on the science behind would be in 'Articles' vignettes.

Finally, I think the package's visualization examples would be even more
engaging with terrestrial locations, as this would highlight the integration
with base map tiles and provide users with examples that directly relate to
typical wildfire analysis scenarios. If there are specific reasons/constrains
for using oceanic locations in the examples it's fine.

##### Functions

In general, there are several functions on which the documentation starts with
papers reference. I would divide this using the `@description` and implementing
`@details`. In the `roxygen` [documentation](https://roxygen2.r-lib.org/articles/rd.html)
states that, after the title:

> The second paragraph is the description: this comes first in the documentation
and should briefly describe what the function does.

> The third and subsequent paragraphs go into the details: this is a (often
long) section that comes after the argument description and should provide any
other important details of how the function operates. The details are optional.

I suggest using the details section to include the reference to a paper if
needed.

There are some functions which have fixed values in their bodies. I'm wondering
if those could be arguments with default values that the user can change. This
is [one example](https://github.com/heyairf/fireexposuR/blob/b689adb450e53090c0698bdcabbb442c5af9915a/R/fire_exp_dir.R#L53)

 - `fire_exp_map_class()` function documentation states: Scales and colors are
 determined by the parameter classify which can be set to "local" or "landscape"
 This are the kind of explanations I would include in a 'article' in the
 vignettes. Going further explaining to a user the differences of some of the
 arguments options in the results and the theory behind it could be part of the
 things to be explained in a 'article' vignette.
