### 2024-11-28 #659 @maurolepore
@maelle unfortunately I'm still seeking the second reviewer. It seems to become harder and harder that as we approach the end of the year. I'm about to ask for help in the editors channel.

@heyairf I'm really sorry for the delay. It's not super common but occasionally it does happen that it takes quite some effort.
