### 2024-12-20 #659 @huizezhang-sherry
- [x]  As the reviewer I confirm that there are no [[conflicts of interest](https://devguide.ropensci.org/policies.html#coi)](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

---

- [ ]  The package would benefit from more documentation on the methods and how outputs are computed. While the papers referenced provide comprehensive details, a good rule of thumb is that the package documentation(README, articles, and function reference) should be self-contained, with publication(s) referenced for further details.
    - [ ]  It would be helpful to explain how the summary in `fire_exp_dir()` is generated and what is plotted in `fire_exp_dir_plot()` to help user comprehend the plot.  For example, it would be beneficial to clarify that `fire_exp_dir()` divides the space around the point into 1-degree intervals and computes the binary variable, `viable` as whether trajectories with at least 80% of its length traversing high exposure(60%). Similarly `fire_exp_dir_plot()` plots those directional segment area where `viable = 1`.
    - [ ]  The basemap example in `fire_exp_dir_plot()` could be more informative if it were draw on terrain rather than a pseudo-example over the ocean. Also, It will be helpful to include technical details about how the plot is created. For example, when the basemap is used (map = TRUE),   `tidyterra::geom_spatvector()/ tidyterra::geom_spatraster_rgb()` are used for plotting. Without the basemap, `geom_bar()` is used after casting the `SpatVector` object to a data frame.
    - [ ]  The author may consider explaining the difference of using `fire_exp_dir_multi()` and `fire_exp() + fire_exp_dir() + fire_exp_dir_plot()`
- [ ]  The use of color (red, orange, and yellow) in `fire_exp_dir_plot()`, while differentiates the three distance intervals (to 5km, to 10km, and to 15km), could be misread as the magnitude of the variable being plotted if users are not clear about the variable. Since this distance information is already captured by their relative position in the ring, the authors may want to reconsider the appropriateness of using color in this context.
- [ ]  `fire_exp_dir()` : authors could consider making the directional interval (currently fixed 1 degree) an argument in `fire_exp_dir()` to allow users to summarize and visualize trajectories at customized finer or coarser intervals. Similarly for segment length (5km apart: 0-5km/ 5-10km/ 10-15km), proportional length traversing (80%), exposure percentage (60%).
- [ ]  Function names in the package could more clearly reflect their purposes. The package provides two types of functions: 1) those that summarize fire exposure metrics and 2) those that plot fire exposure. It would be helpful if function names can explicitly indicates their purpose. For example, standardizing visualization functions as  `fire_exp_vis_xxx()` would make tehm easier to identify. Also authors may consider organizing functions into  sections on the pkgdown reference page.
