### 2024-12-22 #659 @maurolepore
@heyairf you can now go ahead and incorporate both reviews. Thank you for being patient :pray:

https://devdevguide.netlify.app/softwarereview_author#the-review-process

> We ask that you respond to reviewers’ comments within 2 weeks of the last-submitted review, but you may make updates to your package or respond at any time. Your response should include a link to the updated [NEWS.md](https://devdevguide.netlify.app/maintenance_releases#news) of your package. Here is [an author response example](https://github.com/ropensci/software-review/issues/160#issuecomment-355043656). Once the response is commited, [submit it using the bot](https://devdevguide.netlify.app/bot_cheatsheet#submit-response-to-reviewers). We encourage ongoing conversations between authors and reviewers. See the [reviewing guide](https://devdevguide.netlify.app/softwarereview_reviewer) for more details.
