### 2024-12-28 #659 @maurolepore
Absolutely, the end of the year is inconvenient for everyone
