### 2025-01-06 d2aed89
solution for case of no viable transects
