+++ R/fire_exp_dir.R
@@ -1,47 +1,172 @@
 #' Conduct a directional exposure assessment
 #'
-#' @description `fire_exp_dir()` automates the directional vulnerability assessment
-#'   methods from Beverly and Forbes 2023. This function can return the
-#'   assessment transects as a spatial feature of transect segments or a table
-#'   of each transect segment as a data frame
-#'   The exposure layer and value layer must have a CRS defined.
-#'   If a polygon is used, it must be a simple feature without holes or complex
-#'   geometry. The function will not run if the polygon should be simplified
-#'   further. See Forbes and Beverly 2024 (Manuscript in preparation) for data
-#'   preparation advice.
+#' @description `fire_exp_dir()` returns a SpatVector of transects that have a
+#'  binary "Viable" attribute that represents if it is a viable linear pathway
+#'  a wildfire could follow toward a value of interest.
+#'
+#'
+#' @details
+#'  `fire_exp_dir()` automates the directional vulnerability assessment
+#'  methods documented in Beverly and Forbes (2023).
+#'
+#'  ## Spatial Reference
+#'
+#'  The inputs for the exposure and value layer must have the same coordinate
+#'  reference system (CRS) defined. The transects will be returned in the same
+#'  CRS as the inputs.
+#'
+#'  This function draws the transects by calculating the end point of a
+#'  transect by finding the shortest path along the WGS84 (EPSG: 4326)
+#'  ellipsoid at a given bearing and distance. The `value` input is reprojected
+#'  from the input CRS to latitude and longitude using WGS84 for the
+#'  calculations. After the transects are created, they are projected to match
+#'  the CRS of the input exposure and value layer. The lengths of the projected
+#'  transects will be effected by the scale factor of the input CRS; however,
+#'  the geodesic lengths are maintained.
+#'
+#'  ## Value feature
+#'
+#'  The value feature can be provided as a point or a simplified polygon.
+#'
+#'  If using a point feature the analysis can be sensitive to the placement
+#'  of a point. For example, if using a point to represent a large town a
+#'. point placed at the centroid will likely have different outputs than a
+#'  point placed at the edge of the community due to the arrangement of lower
+#'  exposure values typical of a built environment.
+#'
+#'  An option to use a simplified polygon has been added to this function for
+#'  values that may be too large to represent with a single point. The polygon
+#'  should be drawn with the consideration of the following or the function
+#'  will not be able to run. The polygon must be a single part polygon with no
+#'  holes. The polygon should have a smooth and simple shape, ideally circular
+#'  or ellipsoidal. The polygon should also have an approximate radius of less
+#'  than 5000 meters from the center. If the area of interest is larger than
+#'  this then consider using multiple assessments.
+#'
+#'  ## Default Values
+#'  The default values are based on the methods used and validated in Alberta,
+#'  Canada by Beverly and Forbes (2023). Options have been added to the
+#'  function to allow these values to be adjusted by the user if required.
+#'  Adjusting these values can lead to unexpected and uncertain results.
+#'
+#'  ## Adjusting the transects
+#'  The drawing of the transects can be customized by varying the intervals and
[…]
