+++ man/fire_exp_dir.Rd
@@ -4,12 +4,35 @@
 \alias{fire_exp_dir}
 \title{Conduct a directional exposure assessment}
 \usage{
-fire_exp_dir(exposure, value, table = FALSE)
+fire_exp_dir(
+  exposure,
+  value,
+  lengths = c(5000, 5000, 5000),
+  interval = 1,
+  thresh_exp = 0.6,
+  thresh_viable = 0.8,
+  table = FALSE
+)
 }
 \arguments{
 \item{exposure}{SpatRaster (e.g. from \code{\link[=fire_exp]{fire_exp()}})}
 
 \item{value}{Spatvector of value as a point or simplified polygon}
+
+\item{lengths}{(Optional) A vector of three numeric values. The length of
+each transect starting from the value in meters. The default is
+\code{c(5000, 5000, 5000)}.}
+
+\item{interval}{(Optional) Numeric. The degree interval at which to draw
+the transects for analysis. Can be one of 0.5, 1, 2, 3, 4, 5, 6, 8, or 10.
+The default is \code{1}.}
+
+\item{thresh_exp}{(optional) Numeric. The exposure value to use to define
+high exposure as a proportion. Must be between 0-1. The default is \code{0.6}.}
+
+\item{thresh_viable}{(optional) Numeric. The minimum intersection of a
+transect with a high exposure patch to be defined as a viable pathway as
+a proportion. Must be between 0-1. The default is \code{0.8}.}
 
 \item{table}{Boolean, when \code{TRUE}: returns a table instead of a feature. The
 default is \code{FALSE}.}
@@ -22,27 +45,133 @@ viable. Unless:
 }
 }
 \description{
+\code{fire_exp_dir()} returns a SpatVector of transects that have a
+binary "Viable" attribute that represents if it is a viable linear pathway
+a wildfire could follow toward a value of interest.
+}
+\details{
 \code{fire_exp_dir()} automates the directional vulnerability assessment
-methods from Beverly and Forbes 2023. This function can return the
-assessment transects as a spatial feature of transect segments or a table
-of each transect segment as a data frame
-The exposure layer and value layer must have a CRS defined.
-If a polygon is used, it must be a simple feature without holes or complex
-geometry. The function will not run if the polygon should be simplified
-further. See Forbes and Beverly 2024 (Manuscript in preparation) for data
-preparation advice.
+methods documented in Beverly and Forbes (2023).
+\subsection{Spatial Reference}{
+
+The inputs for the exposure and value layer must have the same coordinate
+reference system (CRS) defined. The transects will be returned in the same
+CRS as the inputs.
+
+This function draws the transects by calculating the end point of a
+transect by finding the shortest path along the WGS84 (EPSG: 4326)
+ellipsoid at a given bearing and distance. The \code{value} input is reprojected
+from the input CRS to latitude and longitude using WGS84 for the
+calculations. After the transects are created, they are projected to match
+the CRS of the input exposure and value layer. The lengths of the projected
+transects will be effected by the scale factor of the input CRS; however,
+the geodesic lengths are maintained.
+}
+
+\subsection{Value feature}{
+
+The value feature can be provided as a point or a simplified polygon.
+
+If using a point feature the analysis can be sensitive to the placement
+of a point. For example, if using a point to represent a large town a
+. point placed at the centroid will likely have different outputs than a
+point placed at the edge of the community due to the arrangement of lower
+exposure values typical of a built environment.
+
+An option to use a simplified polygon has been added to this function for
+values that may be too large to represent with a single point. The polygon
+should be drawn with the consideration of the following or the function
+will not be able to run. The polygon must be a single part polygon with no
+holes. The polygon should have a smooth and simple shape, ideally circular
+or ellipsoidal. The polygon should also have an approximate radius of less
[…]
