### 2025-01-09 4a199a3
documentation and example data improvements
