+++ R/fire_exp_dir.R
@@ -1,104 +1,110 @@
 #' Conduct a directional exposure assessment
 #'
-#' @description `fire_exp_dir()` returns a SpatVector of transects that have a
-#'  binary "Viable" attribute that represents if it is a viable linear pathway
-#'  a wildfire could follow toward a value of interest.
+#' @description `fire_exp_dir()` returns a SpatVector of linear transects toward
+#' a value. Transects are assessed as viable wildfire pathways by intersecting
+#' them with areas of high exposure.
 #'
 #'
 #' @details
-#'  `fire_exp_dir()` automates the directional vulnerability assessment
-#'  methods documented in Beverly and Forbes (2023).
+#' `fire_exp_dir()` automates the directional vulnerability assessment
+#' methods documented in Beverly and Forbes (2023).
 #'
-#'  ## Spatial Reference
+#' ## Spatial Reference
 #'
-#'  The inputs for the exposure and value layer must have the same coordinate
-#'  reference system (CRS) defined. The transects will be returned in the same
-#'  CRS as the inputs.
+#' The inputs for the exposure and value layer must have the same coordinate
+#' reference system (CRS) defined. The transects will be returned in the same
+#' CRS as the inputs.
 #'
-#'  This function draws the transects by calculating the end point of a
-#'  transect by finding the shortest path along the WGS84 (EPSG: 4326)
-#'  ellipsoid at a given bearing and distance. The `value` input is reprojected
-#'  from the input CRS to latitude and longitude using WGS84 for the
-#'  calculations. After the transects are created, they are projected to match
-#'  the CRS of the input exposure and value layer. The lengths of the projected
-#'  transects will be effected by the scale factor of the input CRS; however,
-#'  the geodesic lengths are maintained.
+#' This function draws the transects by calculating the end point of a
+#' transect by finding the shortest path along the WGS84
+#' ([EPSG:4326](https://epsg.io/5326)) ellipsoid at a given bearing and
+#' distance. The `value` input is reprojected from the input CRS to latitude and
+#' longitude using WGS84 for the calculations. After the transects are created,
+#' they are projected to match the CRS of the input exposure and value layer.
+#' The lengths of the projected transects will be effected by the scale factor
+#' of the input CRS; however, the geodesic lengths are maintained.
 #'
-#'  ## Value feature
+#' ## Value feature
 #'
-#'  The value feature can be provided as a point or a simplified polygon.
+#' The value feature can be provided as a point or a simplified polygon.
 #'
-#'  If using a point feature the analysis can be sensitive to the placement
-#'  of a point. For example, if using a point to represent a large town a
-#'. point placed at the centroid will likely have different outputs than a
-#'  point placed at the edge of the community due to the arrangement of lower
-#'  exposure values typical of a built environment.
+#' If using a point feature the analysis can be sensitive to the placement
+#' of a point. For example, if using a point to represent a large town a
+#' point placed at the centroid will likely have different outputs than a
+#' point placed at the edge of the community due to the arrangement of lower
+#' exposure values typical of a built environment.
 #'
-#'  An option to use a simplified polygon has been added to this function for
-#'  values that may be too large to represent with a single point. The polygon
-#'  should be drawn with the consideration of the following or the function
-#'  will not be able to run. The polygon must be a single part polygon with no
-#'  holes. The polygon should have a smooth and simple shape, ideally circular
-#'  or ellipsoidal. The polygon should also have an approximate radius of less
-#'  than 5000 meters from the center. If the area of interest is larger than
-#'  this then consider using multiple assessments.
+#' An option to use a simplified polygon has been added to this function for
+#' values that may be too large to represent with a single point. The polygon
[…]
