+++ R/fire_exp_dir_multi.R
@@ -22,19 +22,19 @@
 #' @export
 #'
 #' @examples
-#' #' # read example hazard data ----------------------------------
-#' filepath <- "extdata/hazard.tif"
-#' haz <- terra::rast(system.file(filepath, package = "fireexposuR"))
-#' # example points across the landscape
-#' e <- terra::buffer(terra::vect(terra::ext(haz), crs = haz), -15500)
-#' pts <- terra::spatSample(e, 20)
-#' # ----------------------------------------------------------
+#' # read example hazard data
+#' hazard_file_path <- "extdata/hazard.tif"
+#' hazard <- terra::rast(system.file(hazard_file_path, package = "fireexposuR"))
 #'
-#' exp <- fire_exp(haz, tdist = "l")
-#' # this example will take a while to run
-#' \dontrun{
-#' fire_exp_dir_multi(exp, pts, plot = TRUE)
-#' }
+#' # generate 10 random example points within the hazard extent
+#' e <- terra::buffer(terra::vect(terra::ext(hazard), crs = hazard), -15500)
+#' points <- terra::spatSample(e, 10)
+#'
+#' # compute exposure metric
+#' exposure <- fire_exp(hazard)
+#'
+#' # plot directional load for multiple points
+#' fire_exp_dir_multi(exposure, points, plot = TRUE)
 
 fire_exp_dir_multi <- function(exposure, values, plot = FALSE, all = FALSE) {
   stopifnot("`exposure` must be a SpatRaster object"
@@ -49,14 +49,6 @@ fire_exp_dir_multi <- function(exposure, values, plot = FALSE, all = FALSE) {
   expl <- exposure
   fts <- values
 
-  # test that values features are within exposure extent before running
-  buff <- terra::buffer(fts, 15000) %>%
-    terra::aggregate()
-  explext <- terra::classify(expl, c(-Inf,Inf,1)) %>%
-    terra::as.polygons()
-  stopifnot("Values features must be within extent of the exposure layer"
-            = terra::relate(buff, explext, "coveredby") == TRUE)
-
   df <- data.frame()
 
   for (i in seq_len(nrow(fts))) {
@@ -69,9 +61,9 @@ fire_exp_dir_multi <- function(exposure, values, plot = FALSE, all = FALSE) {
   }
 
   df2 <- df %>%
-    dplyr::mutate(full = ifelse(.data$to5 + .data$to10 + .data$to15 == 3,
+    dplyr::mutate(full = ifelse(.data$seg1 + .data$seg2 + .data$seg3 == 3,
                                 1, 0)) %>%
-    dplyr::mutate(outer = ifelse(.data$to10 + .data$to15 == 2, 1, 0))
+    dplyr::mutate(outer = ifelse(.data$seg2 + .data$seg3 == 2, 1, 0))
 
   if (plot == TRUE) {
     dfsums <- df2 %>%

[…]
