+++ NAMESPACE
@@ -4,6 +4,7 @@ export("%>%")
 export(fire_exp)
 export(fire_exp_adjust)
 export(fire_exp_dir)
+export(fire_exp_dir_map)
 export(fire_exp_dir_multi)
 export(fire_exp_dir_plot)
 export(fire_exp_extract)
