### 2025-01-09 b26d554
WIP. vignette and pkgdown site improvements
