+++ vignettes/communityscale.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "communityscale"
+title: "Local scale example"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{communityscale}
@@ -35,44 +35,50 @@ library(fireexposuR)
 
 ### Data preperation
 
-First, some example data will be generated for an imaginary town:
+First, some example data will be generated for a town that doesn't exist in the
+middle of Alberta.
 
 ```{r data}
+# load terra for spatial data funtions
 library(terra)
-# read example hazard data ----------------------------------
-filepath <- "extdata/hazard.tif"
-haz <- terra::rast(system.file(filepath, package = "fireexposuR"))
-# read example AOI
-filepath <- "extdata/builtsimpleexamplegeom.csv"
-g <- read.csv(system.file(filepath, package = "fireexposuR"))
-aoi <- terra::vect(as.matrix(g), "polygons", crs = haz)
-# generate random points
-pts <- terra::spatSample(aoi, 200)
-#' # ----------------------------------------------------------
+
+# read example hazard data
+hazard_file_path <- "extdata/hazard.tif"
+hazard <- terra::rast(system.file(hazard_file_path, package = "fireexposuR"))
+
+# read example polygon geometry for area of interest boundary
+geom_file_path <- "extdata/builtsimpleexamplegeom.csv"
+geom <- read.csv(system.file(geom_file_path, package = "fireexposuR"))
+
+# use geometry to make an area of interest polygon
+aoi <- terra::vect(as.matrix(geom), "polygons", crs = hazard)
+
+# generate random points within the aoi to represent values
+points <- terra::spatSample(aoi, 200)
 ```
 
-This example data represents the sort of data that would be pre-prepared
-by a user for the assessment.
 
 #### Hazard data
 
-The `haz` layer is a binary raster that represents wildland fuels that
+The `hazard` layer is a binary raster that represents wildland fuels that
 are able to generate long-range embers up to a transmission distance of
 500 meters:
 
 ```{r hazardvis, echo = FALSE}
-plot(haz)
+# plot hazard layer with terra defaults
+plot(hazard)
 ```
 
 #### Area of interest and values data
 
-The `aoi` layer is a polygon representing the built environment of our
-imaginary town. The `pts` feature represents the centroids of homes and
-structures within the community.
+The `aoi` layer is a polygon representing the boundary of a built environment of
+this imaginary town. The `points` feature represents values within the
+community. We will pretend the points are the locations of structures.
 
 ```{r aoivis, echo = FALSE}
+# plot the vector data with terra defaults
 plot(aoi)
-plot(pts, add = T)
+plot(points, add = T)
 ```
 
 ### Compute exposure
@@ -82,7 +88,7 @@ ember transmission.
 
 ```{r computeexposure}
 library(fireexposuR)
-exp <- fire_exp(haz, tdist = "l")
+exposure <- fire_exp(hazard, tdist = "l")
 ```
 
 ### Visualize exposure
@@ -92,18 +98,16 @@ used to visualize it in different ways. For a landscape, we can map
 exposure with a continuous scale with `fire_exp_map_cont()`:
 
 ```{r maplandscape}
-fire_exp_map_cont(exp)
+fire_exp_map_cont(exposure)
 ```
 
 We can also see how that exposure is distributed within the built
 environment with exposure classifications in an area of interest with
 `fire_exp_map_class()`.
 
-*Note: our imaginary town is in the middle of the Pacific Ocean so the
-base map does not provide further reference.*
 
 ```{r maplocal}
-fire_exp_map_class(exp, classify = "local", aoi)
+fire_exp_map_class(exposure, classify = "local", aoi)
 ```
 
 This map gives us a better understanding of areas of the town that could
[…]
