+++ _pkgdown.yml
@@ -2,3 +2,24 @@ url: https://heyairf.github.io/fireexposuR/
 template:
   bootstrap: 5
 
+reference:
+- title: "Compute"
+  desc: >
+    Functions for automating wildfire exposure methods
+  contents:
+  - fire_exp
+  - fire_exp_adjust
+  - fire_exp_dir
+  - fire_exp_extract
+  - fire_exp_validate
+- title: "Visualize"
+  desc: >
+    Functions for visualizing wildfire exposure outputs in R
+  contents:
+  - fire_exp_summary
+  - fire_exp_map_class
+  - fire_exp_map_cont
+  - fire_exp_extract_vis
+  - fire_exp_dir_map
+  - fire_exp_dir_plot
+  - fire_exp_dir_multi
