### 2025-01-09 2f97a10
documentation and example data readability improvements
