+++ R/fire_exp_adjust.R
@@ -1,6 +1,12 @@
 #' Compute the wildfire exposure metric with custom transmission distances
 #'
-#' @description For advanced users. If the transmission distances from the
+#' @description For advanced users. Adjust the transmission distances from the
+#' defaults used in [fire_exp()].
+#'
+#' @details
+#' **DOCUMENTATION IN DEVELOPMENT**
+#'
+#' If the transmission distances from the
 #' wildfire exposure literature are not representative of the wildland fuels
 #' in your area of interest, this function can be used to change the
 #' transmission distance to a custom distance. It is highly recommended that
@@ -22,13 +28,12 @@
 #' @export
 #'
 #' @examples
-#' # read example hazard data ----------------------------------
-#' filepath <- "extdata/hazard.tif"
-#' haz <- terra::rast(system.file(filepath, package = "fireexposuR"))
-#' # -----------------------------------------------------------
+#' # read example hazard data
+#' hazard_file_path <- "extdata/hazard.tif"
+#' hazard <- terra::rast(system.file(hazard_file_path, package = "fireexposuR"))
 #'
 #' # compute long range exposure with custom disance of 800 m
-#' exp <- fire_exp_adjust(haz, tdist = 800)
+#' exposure800 <- fire_exp_adjust(hazard, tdist = 800)
 #'
 fire_exp_adjust <- function(hazard, tdist, nonburnable) {
   stopifnot("`hazard` must be a SpatRaster object"
