+++ tests/testthat/test-fire_exp_dir_map.R
@@ -1,3 +1,48 @@
-test_that("multiplication works", {
-  expect_equal(2 * 2, 4)
+# repeat test data
+filepath <- "extdata/hazard.tif"
+haz <- terra::rast(system.file(filepath, package = "fireexposuR"))
+filepath <- "extdata/polygon_geometry.csv"
+g <- read.csv(system.file(filepath, package = "fireexposuR"))
+v <- terra::vect(as.matrix(g), "polygons", crs = haz)
+nb <- terra::rasterize(v, haz)
+pts <- terra::spatSample(v, 20)
+pt_wkt <- "POINT (400000 6050000)"
+pt <- terra::vect(point_wkt, crs = hazard)
+
+
+exp <- fire_exp(haz)
+
+t_pt <- fire_exp_dir(exp, pt)
+t_pol <- fire_exp_dir(exp, v)
+
+
+# tests ========================================================================
+
+test_that("fire_exp_dir_map() input checks work", {
+  expect_error(fire_exp_dir_map(2),
+               "`transects` must be a SpatVector object")
+  expect_error(fire_exp_dir_map(t_pt, title = 2),
+               "`title` must be")
+  expect_error(fire_exp_dir_map(t_pt, labels = "blah"),
+               "`labels` must be") # not enough
+  expect_error(fire_exp_dir_map(t_pt, labels = c("blah", "blah")),
+               "`labels` must be") # not characters
+  expect_error(suppressMessages(fire_exp_dir_map(t_pt, value = 2)),
+               "`value` must be a SpatVector object")
+})
+
+test_that("fire_exp_dir_map() returns objects with correct class", {
+  expect_s3_class(suppressMessages(fire_exp_dir_map(t_pt)), "ggplot")
+})
+
+test_that("fire_exp_dir_map() runs when input conditions are met", {
+  expect_no_error(suppressMessages(fire_exp_dir_map(t_pt)))
+  expect_no_error(suppressMessages(fire_exp_dir_map(t_pol)))
+  expect_no_error(suppressMessages(fire_exp_dir_map(t_pol, value = v)))
+  expect_no_error(suppressMessages(fire_exp_dir_map(t_pt,
+                                                    labels = c("blah",
+                                                               "blah",
+                                                               "blah"))))
+  expect_no_error(suppressMessages(fire_exp_dir_map(t_pt,
+                                                    title = "blah blah blah")))
 })
