### 2025-01-10 e1b92e1
elliminate fixed value
