### 2025-01-10 #659 @heyairf
Hi everyone,

I just wanted to say that your comments and suggestions have been so helpful in thinking about what the documentation should look like. As I have been going through the package to incorporate your feedback to all the current functions I have realized it is quite a monumental task to go through it all and it is taking me a considerable amount of time. I have pushed a few changes to the main branch to respond to your comments with examples of the changes I am making, but I have not completed documenting all the functions yet. I am also drafting a better README file to incorporate some more of your great feedback.

I have marked my responses with numbers to make them easier to respond to if required. I don’t expect you to take the time to reply until I am finished unless you want to! I will let you know as soon as that is. The only comment I would appreciate a response to if possible is number 4… if any of you could help me identify where I went wrong with that I’d be grateful.

Thanks,
Air

Response to @ronnyhdez

_It would be helpful to clarify how coordinate reference systems are handled throughout the package, particularly whether transformations affect only the visualization outputs or also the underlying analyses. For example, the use of tiles involves the re-projection to EPSG:3857, which is necessary for visualization. Given that the target audience of this package seems to be wildfire scientists, documenting these spatial operations would enhance the functions' usability._

1.  This is a great suggestion. I have added spatial reference descriptions to the details of these functions [here](https://github.com/heyairf/fireexposuR/blob/e1b92e1d60146dcab301cd73db1632f5f90388f0/R/fire_exp_dir.R#L12) and [here](https://github.com/heyairf/fireexposuR/blob/e1b92e1d60146dcab301cd73db1632f5f90388f0/R/fire_exp_dir_map.R#L12) (and will add the same to all other functions that have transformations as I continue populating the documentation).

_Also, it would be helpful to indicate the geographical scope of the implemented methods. From the references included in the README, it seems that the studies were all conducted in Alberta, Canada. If the methods are specifically calibrated for this region's characteristics, this should be explicitly stated in the package documentation to help users assess the applicability to their study areas._

2. This is a great point. I will be adding this information to the README as I continue with editing the documentation.

_While references to scientific publications provide valuable methodological context, the function documentation would be more effective if it primarily focused on usage and implementation details, with citations serving as supplementary information. This would enhance the user experience by providing immediate access to practical implementation guidance._

3. You are so right. The inspiration behind this project was to make the methods more accessible, and this could definitely be a barrier. Updates to the documentation will reflect this as well.

_I would suggest adding a CITATION file (see rOpenSci guidelines[here](https://devguide.ropensci.org/pkg_building.html#citation-file)) to help users properly acknowledge your work when using the package in their research. The checkpoiints above state that there should be a 'contribution guidelines'.
[Here](https://devguide.ropensci.org/maintenance_collaboration.html#contributing-guide) is some information that can guide you about how to include it._

4. I believe that I have included both of these files? The info in them is drawn on the pkgdown site… but please advise if the files need to be moved or linked somehow in the files.

_Specific thoughts
README and Vignettes (abridged)_

5.  This feedback is so helpful and will be incorporated into a new README file. The community scale example is also being updated. I am also drafting a page (article or vignette?) that will offer a brief description of the input data requirements.

_code snippets_

6. Great suggestion. I have made changes to all of the examples throughout the package to increase the readability. Here's an [example](https://github.com/heyairf/fireexposuR/blob/e1b92e1d60146dcab301cd73db1632f5f90388f0/R/fire_exp_dir_map.R#L40)

_Finally, I think the package's visualization examples would be even more engaging with terrestrial locations, as this would highlight the integration with base map tiles and provide users with examples that directly relate to typical wildfire analysis scenarios. If there are specific reasons/constraints for using oceanic locations in the examples it's fine._

7.  I have updated the hazard raster in the example data folder to somewhere in the middle of AB! This is the same data that is used in the function examples as well. You can see the visualization at the end of this [vignette](https://heyairf.github.io/fireexposuR/articles/communityscale.html)

_Functions
In general, there are several functions on which the documentation starts with papers reference. I would divide this using the `@description` and implementing `@details`. In the roxygen [documentation](https://roxygen2.r-lib.org/articles/rd.html)
states that, after the title: The second paragraph is the description: this comes first in the documentation and should briefly describe what the function does. The third and subsequent paragraphs go into the details: this is a (often long) section that comes after the argument description and should provide any other important details of how the function operates. The details are optional. I suggest using the details section to include the reference to a paper if needed._

8. Thank you for this feedback! I have begun to implement it in all of the function documentation. I have shorted the descriptions of all the functions and have made them more descriptive. So far, the populated details sections have only been pushed for [these](https://github.com/heyairf/fireexposuR/blob/e1b92e1d60146dcab301cd73db1632f5f90388f0/R/fire_exp_dir.R) [three](https://github.com/heyairf/fireexposuR/blob/e1b92e1d60146dcab301cd73db1632f5f90388f0/R/fire_exp_dir_map.R) [functions](https://github.com/heyairf/fireexposuR/blob/e1b92e1d60146dcab301cd73db1632f5f90388f0/R/fire_exp_dir_plot.R).

_There are some functions which have fixed values in their bodies. I'm wondering if those could be arguments with default values that the user can change. This is [one example](https://github.com/heyairf/fireexposuR/blob/b689adb450e53090c0698bdcabbb442c5af9915a/R/fire_exp_dir.R#L53)_

9.  I have added more options and parameter adjustments to the three functions linked above in comment 8 to add more options for customization and will continue on this trend with the other functions, including the option for custom classification breaks as I continue with the updates!

_`fire_exp_map_class()` function documentation states: Scales and colors are determined by the parameter classify which can be set to "local" or "landscape". This are the kind of explanations I would include in a 'article' in the vignettes. Going further explaining to a user the differences of some of the arguments options in the results and the theory behind it could be part of the things to be explained in a 'article' vignette._

10.  Would it be appropriate to have that explanation in the roxygen details for now? This would be a great topic to write an article about in the future for sure.

Response to @huizezhang-sherry

_The package would benefit from more documentation on the methods and how outputs are computed. While the papers referenced provide comprehensive details, a good rule of thumb is that the package documentation(README, articles, and function reference) should be self-contained, with publication(s) referenced for further details._

11.  I agree! I am actively working on this but have only pushed updates for [these](https://github.com/heyairf/fireexposuR/blob/e1b92e1d60146dcab301cd73db1632f5f90388f0/R/fire_exp_dir.R) [three](https://github.com/heyairf/fireexposuR/blob/e1b92e1d60146dcab301cd73db1632f5f90388f0/R/fire_exp_dir_map.R) [functions](https://github.com/heyairf/fireexposuR/blob/e1b92e1d60146dcab301cd73db1632f5f90388f0/R/fire_exp_dir_plot.R).

_It would be helpful to explain how the summary in `fire_exp_dir()` is generated and what is plotted in `fire_exp_dir_plot()` to help user comprehend the plot. For example, it would be beneficial to clarify that `fire_exp_dir()` divides the space around the point into 1-degree intervals and computes the binary variable, viable as whether trajectories with at least 80% of its length traversing high exposure(60%). Similarly `fire_exp_dir_plot()` plots those directional segment area where `viable = 1`._

12. I have made expansions to the documentation for these functions (above). I have also split the visualization into two separate functions `fire_exp_dir_plot()` which plots a radial graph, and `fire_exp_dir_map()` which maps the transects over a map tile and described them in the details sections.

_The basemap example in `fire_exp_dir_plot()` could be more informative if it were draw on terrain rather than a pseudo-example over the ocean. Also, It will be helpful to include technical details about how the plot is created. For example, when the basemap is used `(map = TRUE)`, `tidyterra::geom_spatvector()`/ `tidyterra::geom_spatraster_rgb()` are used for plotting. Without the basemap, `geom_bar()` is used after casting the SpatVector object to a data frame._

13. Same as the above linked functions. The technical details about how the two functions directional visualizations are produced have been added to the function documentation.

_The author may consider explaining the difference of using `fire_exp_dir_multi()` and `fire_exp()` + `fire_exp_dir()` + `fire_exp_dir_plot()`_

14.  Great suggestion, I will be incorporating this into the details of `fire_exp_dir_multi()` as I continue to populate the documentation.

_The use of color (red, orange, and yellow) in `fire_exp_dir_plot()`, while differentiates the three distance intervals (to 5km, to 10km, and to 15km), could be misread as the magnitude of the variable being plotted if users are not clear about the variable. Since this distance information is already captured by their relative position in the ring, the authors may want to reconsider the appropriateness of using color in this context._

15.  The colours were chosen to mirror the colours used in Beverly and Forbes(2023), and do aim to represent a level of magnitude in terms of wildfire exposure to the value(centre) as a pathway far away (yellow) is less concerning than a pathway close by(red)

_`fire_exp_dir()` : authors could consider making the directional interval (currently fixed 1 degree) an argument in `fire_exp_dir()` to allow users to summarize and visualize trajectories at customized finer or coarser intervals. Similarly for segment length (5km apart: 0-5km/ 5-10km/ 10-15km), proportional length traversing (80%), exposure percentage (60%)._

16. This is a great suggestion! I have added the option to change all four of these parameters to `fire_exp_dir()` to customize the analysis. I have also added details into the documentation to describe what changing them will do. I will continue to reduce the hard coded values in the other functions to increase the customization options!

_Function names in the package could more clearly reflect their purposes. The package provides two types of functions: 1) those that summarize fire exposure metrics and 2) those that plot fire exposure. It would be helpful if function names can explicitly indicates their purpose. For example, standardizing visualization functions as `fire_exp_vis_xxx()` would make tehm easier to identify. Also authors may consider organizing functions into sections on the pkgdown reference page._

17.  have organized the pkgdown reference page into two categories: Compute and Visualize. The compute functions will do some kind of analysis and return spatial data, where the visualize functions return either a graphic or a table. I will consider how the function naming could be improved. I am slightly hesitant to change the function names just yet, so I hope that the reference organization helps for now.

In summary:

I’m still working on things but I am dedicating lots of time to this! If you have comments you’d like to respond to now please do. I have no problems with you waiting until I have fully incorporated your feedback into a final polished product. Thanks so much for all of your helpful feedback.  :)
