### 2025-01-22 ac7381b
added input check for users who made exposure raster in other programs
