+++ tests/testthat/test-fire_exp_dir_multi.R
@@ -25,6 +25,6 @@ test_that("fire_exp_dir_multi() input checks and function messages work", {
 })
 
 test_that("fire_exp_dir_multi() runs when input conditions are met", {
-  expect_no_error(fire_exp_dir_multi(exp, pts, plot = T))
-  expect_no_error(fire_exp_dir_multi(exp, pts))
+  expect_no_error(fire_exp_dir_multi(exp, pts, plot = T, interval = 10))
+  expect_no_error(fire_exp_dir_multi(exp, pts, interval = 10))
 })
