+++ R/fire_exp_summary.R
@@ -1,33 +1,43 @@
 #' Summarize exposure by class
 #'
 #' @description `fire_exp_summary()` creates a summary table of area and
-#' proportions of exposure in predetermined exposure classes.
+#' proportions of exposure in predetermined or custom exposure classes.
 #'
 #' @details
-#' **DOCUMENTATION IN DEVELOPMENT**
 #'
-#' The table reports the number of pixels, the
-#' proportion, and area in hectares and meters squared by class.
+#' This function summarizes the outputs from [fire_exp()] with classes.
+#' Classes can be chosen from the pre-set `"local"` and `"landscape"` options,
+#' or customized. To use a custom classification scheme, it should be defined
+#' with a list of numeric vectors defining the upper limits of the breaks. A
+#' Nil class is added automatically for exposure values of exactly zero.
 #'
-#' Landscape classification breaks are:
-#' * Low (0-20%)
-#' * Moderate (20-40%)
-#' * High (40-60%),
-#' * Very High (60-80%)
-#' * Extreme (80-100%)
+#' Local classification breaks are predefined as `c(0.15, 0.3, 0.45, 1)`:
+#' * Nil (0)
+#' * 0 - 0.15
+#' * 0.15 - 0.3
+#' * 0.3 - 0.45
+#' * 0.45 - 1
 #'
-#' Local classification breaks are:
-#' * Nil (0%)
-#' * Low (>0-15%)
-#' * Moderate (15-30%)
-#' * High (30-45%)
-#' * Extreme (45%+)
+#' Landscape classification breaks are predefined as `c(0.2, 0.4, 0.6, 0.8, 1)`:
+#' * Nil (0)
+#' * 0 - 0.2
+#' * 0.2 - 0.4
+#' * 0.4 - 0.6
+#' * 0.6 - 0.8
+#' * 0.8 - 1
+#'
+#' The table reports the number of pixels, the proportion, and area in hectares
+#' and meters squared in each class.
+#'
 #'
 #' @param exposure SpatRaster from [fire_exp()]
 #' @param aoi (optional) SpatVector of an area of interest to mask exposure for
-#'   summary
-#' @param classify character, either `"landscape"` or `"local"`. default is
-#'   `"landscape"`.
+#' summary
+#' @param classify character, either `"local"`, `"landscape"`, or `"custom"`,
+#' to specify classification scheme to use. The default is `"local"`.
+#' If set to `"custom"`: the parameter `class_breaks` must be used.
+#' @param class_breaks vector of numeric values between 0-1 of the upper limits
+#' of each custom class. Ignored unless `classify = "custom"`. See details.
 #'
 #' @returns a summary table as a data frame object
 #' @export
@@ -52,11 +62,13 @@
 #' fire_exp_summary(exposure, aoi, classify = "landscape")
 #'
 fire_exp_summary <- function(exposure, aoi,
-                             classify = c("landscape", "local")) {
+                             classify = c("landscape", "local", "custom"),
+                             class_breaks) {
   stopifnot("`exposure` must be a SpatRaster object"
             = class(exposure) == "SpatRaster")
   stopifnot("`exposure` layer must have values between 0-1"
-            = (round(terra::minmax(exposure)[1], 0) >= 0 && round(terra::minmax(exposure)[2], 0) <= 1))
+            = (round(terra::minmax(exposure)[1], 0) >= 0
+               && round(terra::minmax(exposure)[2], 0) <= 1))
   stopifnot("Linear units of `exposure` layer must be in meters"
[…]
