+++ R/fire_exp_map_class.R
@@ -5,42 +5,62 @@
 #'
 #'
 #' @details
-#' **DOCUMENTATION IN DEVELOPMENT**
-#' This function dynamically pulls map tiles for a base map,
-#' so it is recommended the area of interest is localized. The zoom level may
-#' need to be adjusted based on the extent of your data; see
+#'
+#' This function returns a standardized map with basic cartographic elements.
+#'
+#' The plot is returned as a ggplot object which can be exported/saved to
+#' multiple image file formats.
+#'
+#' This function visualizes the outputs from [fire_exp()] with classes.
+#' Classes can be chosen from the pre-set `"local"` and `"landscape"` options,
+#' or customized. To use a custom classification scheme, it should be defined
+#' with a list of numeric vectors defining the upper limits of the breaks. A
+#' Nil class is added automatically for exposure values of exactly zero.
+#'
+#' Local classification breaks are predefined as `c(0.15, 0.3, 0.45, 1)`:
+#' * Nil (0)
+#' * 0 - 0.15
+#' * 0.15 - 0.3
+#' * 0.3 - 0.45
+#' * 0.45 - 1
+#'
+#' Landscape classification breaks are predefined as `c(0.2, 0.4, 0.6, 0.8, 1)`:
+#' * Nil (0)
+#' * 0 - 0.2
+#' * 0.2 - 0.4
+#' * 0.4 - 0.6
+#' * 0.6 - 0.8
+#' * 0.8 - 1
+#'
+#' ## Spatial reference
+#'
+#' This function dynamically pulls map tiles for a base map.
+#' The inputs are projected to WGS 84/Pseudo-Mercator
+#' ([EPSG:3857](https://epsg.io/3857)) to align them with the map tiles.
+#'
+#' ## Zoom level
+#' The map tile zoom level may need to be adjusted. If the base map is blurry,
+#' increase the zoom level. Higher zoom levels will slow down the function, so
+#' only increase if necessary. Reference the
 #' [OpenStreetMap Wiki](https://wiki.openstreetmap.org/wiki/Zoom_levels) for
 #' more information on zoom levels.
-#' For mapping large extents it is recommended (and will be faster) to use
-#' [fire_exp_map_cont()] which does not use base maps.
 #'
-#' Scales and colors are determined by the parameter `classify`
-#' which can be set to `"local"` or `"landscape"`.
-#'
-#' Landscape classification breaks are:
-#' * Low (0-20%)
-#' * Moderate (20-40%)
-#' * High (40-60%),
-#' * Very High (60-80%)
-#' * Extreme (80-100%)
-#'
-#' Local classification breaks are:
-#' * Nil (0%)
-#' * Low (>0-15%)
-#' * Moderate (15-30%)
-#' * High (30-45%)
-#' * Extreme (45%+)
-#'
-#'
+#' @seealso [fire_exp_map_class()]
 #'
 #' @param exposure SpatRaster (e.g. from [fire_exp()])
-#' @param classify character, either `"local"` or `"landscape"` to specify
-#'   classification scheme to use.
-#' @param aoi SpatVector of an area of interest to mask exposure
-#' @param zoom (Optional). numeric, set the zoom level for the basemap based
-#' on the extent of your data if defaults are not appropriate. Defaults if:
-#' * `classify = "local"` the zoom level default is `13`
-#' * `classify = "landscape"` the zoom level default is `6`
+#' @param aoi (Optional) SpatVector of an area of interest to mask exposure
+#' @param classify character, either `"local"`, `"landscape"`, or `"custom"`,
+#' to specify classification scheme to use. The default is `"local"`.
+#' If set to `"custom"`: the parameter `class_breaks` must be used.
[…]
