### 2025-01-30 8566581
add auto sort of class break values
