+++ R/fire_exp_map_class.R
@@ -135,6 +135,8 @@ fire_exp_map_class <- function(exposure, aoi, classify = c("local", "landscape",
 
   zoom_level <- ifelse(missing(zoom_level), 12, zoom_level)
 
+  class_breaks <- sort(class_breaks)
+
   # class_breaks checks
   stopifnot("`class_breaks` must be a vector of numbers"
             = class(class_breaks) == "numeric")
