+++ README.md
@@ -26,22 +26,30 @@ temporal horizons and spatial extents.
 > If Prometheus has *ever* been installed on your device you must take
 > some additional steps before loading this package (and most other R
 > packages that manipulate spatial data). Please refer to the vignette
-> “Prometheus users” before continuing. **This step is necessary even if
-> you have since uninstalled the program from your computer.**
+> `vignette("prometheus")` before continuing. **This step is necessary
+> even if you have since uninstalled the program from your computer.**
 
 ## Package overview
 
-- provide an accessible platform for conducting wildfire exposure
+- provides an accessible platform for conducting wildfire exposure
   assessments
-- automate methods previously documented in a series of scientific
+- automates methods previously documented in a series of scientific
   publications
-- provide options for customization and validation to meet the needs of
+- provides options for customization and validation to meet the needs of
   users applying the package across diverse use cases and geographic
   areas
 
 ### Who the package is for
 
-gfad
+This package is for anyone who is interested in conducting wildfire
+exposure assessments. This can include, but is not limited to,
+researchers, government agencies, forest industry, consultants,
+communities, and interested individuals.
+
+User’s without pre-existing knowledge of wildland fuels, spatial data,
+and R may need to dedicate more time and effort to learning how to apply
+the tools in this package by exploring the included documentation and
+linked resources.
 
 ### What the package does
 
@@ -64,7 +72,16 @@ package and the plethora of free R tutorials available online.
 
 3.  Visualizes outputs
 
+The outputs from the analysis functions can be quickly visualized in R
+with a selection of visualization functions that return plots, maps, or
+tables which can be exported as images or .csv files.
+
 4.  Allows customization
+
+Functions allow for custom parameterization to alter the analysis if
+desired. This allows for easy adaptation of the methods to suit the
+unique requirements for different use cases, geographic areas, and
+scales of analysis.
 
 ### What the package doesn’t
 
@@ -74,9 +91,9 @@ Effort has been made to provide as much technical detail in the
 documentation of this package; However, it may still be necessary to
 spend some time reading the associated scientific publications that the
 functions in this package are automating to grasp the theory behind the
-methodologies. Citations have been provided with linked DOI’s throughout
+methodologies. Citations have been provided with DOI links throughout
 this manual. This is especially relevant to users who intend to adjust
-the parameter settings from the defaults.
+function parameters.
 
 2.  Prepare your input data
 
@@ -85,20 +102,21 @@ scales. This means that the input data requirements are dependent on the
 intended purpose. The user is responsible for preparing the input data
 before using this package. Input data can be prepared in any geographic
 information system (GIS) program, or done directly in R. Refer to the
-vignette “Preparing input data” for guidance and examples.
+vignette `vignette("prep-input-data")` for guidance and examples.
 
-3.  Decide your custom parameters
+3.  Decide custom parameters
+
+The documentation and resources available in this package aim to help
+inform decisions around custom parameterization, but these decisions
+must be made by the user. Additional analysis may be required to justify
+deviation from the defaults.
 
 4.  Quality check your outputs
 
-**DOCUMENTATION IN DEVELOPMENT**
-
-The goal of fireexposuR is to provide a standardized and accessible
[…]
