### 2025-01-31 09d9b0f
Documentation expansion and improvements
