+++ R/fire_exp_dir_multi.R
@@ -19,7 +19,7 @@
 #' value when a point is used. Adjusting the parameters for `fire_exp_dir()` is
 #' also supported. See details in [`fire_exp_dir()`] for more information.
 #'
-#' ## References
+#' @references
 #' Beverly JL, Forbes AM (2023) Assessing directional vulnerability to
 #' wildfire. *Natural Hazards* **117**, 831-849.
 #' [DOI](https://doi.org/10.1007/s11069-023-05885-3)
