+++ R/fire_exp_dir_plot.R
@@ -13,7 +13,7 @@
 #' The plot is returned as a ggplot object which can be exported/saved to
 #' multiple image file formats.
 #'
-#' ## References
+#' @references
 #' Beverly JL, Forbes AM (2023) Assessing directional vulnerability to
 #' wildfire. *Natural Hazards* **117**, 831-849.
 #' [DOI](https://doi.org/10.1007/s11069-023-05885-3)
