+++ pkgdown.yml
@@ -2,8 +2,10 @@ pandoc: 3.1.11
 pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
-  communityscale: communityscale.html
-last_built: 2025-01-10T19:04Z
+  fireexposuR: fireexposuR.html
+  prep-input-data: prep-input-data.html
+  prometheus: prometheus.html
+last_built: 2025-01-31T03:56Z
 urls:
   reference: https://heyairf.github.io/fireexposuR/reference
   article: https://heyairf.github.io/fireexposuR/articles
