+++ NEWS.md
@@ -1,0 +1,3 @@
+# fireexposuR 1.0.1
+
+* Initial CRAN submission.
