+++ vignettes/resources.Rmd
@@ -1,0 +1,52 @@
+---
+title: "resources"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{resources}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+# Resources
+
+## Further reading and resources
+
+Wildfire exposure research news and course offering announcements are posted on the Wildfire Analytics website
+
+* [wildfireanalytics.org](https://wildfireanalytics.org)
+
+This guidebook provides a basic and easy to read overview of wildfire exposure assessments with very helpful graphics
+
+* FireSmart Canada (2018) Wildfire exposure assessment guidebook. Available [here](https://firesmartcanada.ca/wp-content/uploads/2022/01/FS_ExposureAssessment_Sept2018-1.pdf)
+
+This paper introduced key concepts about wildfire exposure and transmission distances. The guidebook above is based on this research product. Please note that this is the only resource behind a paywall.
+ 
+* Beverly JL, Bothwell P, Conner JCR, Herd EPK (2010) Assessing the exposure of the built environment to potential ignition sources generated from vegetative fuel. *International Journal of Wildland Fire* **19**, 299-313. [DOI](https://doi.org/10.1071/WF09071)
+
+This paper validated the wildfire exposure metric at a landscape scale with observed fire history in the province of Alberta, Canada. Different methods for defining hazardous fuels, and ember spread distances are explored. Results are also compared to outputs from other wildfire models.
+ 
+* Beverly JL, McLoughlin N, Chapman E (2021) A simple metric of landscape fire exposure. *Landscape Ecology* **36**, 785-801. [DOI](https://doi.org/10.1007/s10980-020-01173-8)
+
+This paper introduces the directional exposure assessment methodology. It explains why the default values were chosen for the assessment.
+ 
+* Beverly JL, Forbes AM (2023) Assessing directional vulnerability to wildfire. *Natural Hazards* **117**, 831-849. [DOI](https://doi.org/10.1007/s11069-023-05885-3)
+
+This paper investigates the sensitivity of wildfire exposure assessments to different land cover information products used to create the input hazardous fuel data at local scales
+
+* Forbes AM, Beverly JL (2024) Influence of fuel data assumptions on wildfire exposure assessment of the built environment. *International Journal of Wildland Fire* **33**, WF24025 [DOI](https://doi.org/10.1071/WF24025)
+
+This paper applied directional vulnerability assessments with evacuation planning methods
+
+* Kim AM, Beverly JL, Al Zahid A (2024) Directional analysis of community wildfire evacuation capabilities. *Safety Science* **171**, 106378. [DOI](https://doi.org/10.1016/j.ssci.2023.106378)
+
+This paper applied wildfire exposure assessments in Alaska, USA using alternative methods to define hazardous fuels
+
+* Schmidt JI, Ziel RH, Calef MP, Varvak A (2024) Spatial distribution of wildfire threat in the far north: exposure assessment in boreal communities. *Natural Hazards* **120**, 4901-4924. [DOI](https://doi.org/10.1007/s11069-023-06365-4)
+
