+++ README.Rmd
@@ -41,31 +41,43 @@ temporal horizons and spatial extents.
 > If Prometheus has *ever* been installed on your device you must take
 > some additional steps before loading this package (and most other R
 > packages that manipulate spatial data). Please go through the steps in
-> `vignette("prometheus")` before continuing. **This is necessary even if
-> you have since uninstalled the program from your computer.**
+> `vignette("prometheus")` even if you have since uninstalled the
+> program from your computer.
 
-## Package overview
+## Package Overview
 
--   provides an accessible platform for conducting wildfire exposure assessments
--   automates methods previously documented in a series of scientific publications
--   provides options for customization and validation to meet the needs of users applying the package across diverse use cases and geographic areas
+### Highlights
+
+-   provides an accessible platform for conducting wildfire exposure
+    assessments
+-   automates methods previously documented in a series of scientific
+    publications
+-   provides options for customization and validation to meet the needs
+    of users applying the package across diverse use cases and
+    geographic areas
 
 ### Who the package is for
 
-This package is for anyone who is interested in conducting wildfire exposure assessments. This can include, but is not limited to, researchers, government agencies, forest industry, consultants, communities, and interested individuals.
+This package is for anyone who is interested in conducting wildfire
+exposure assessments. This can include, but is not limited to,
+researchers, government agencies, forest industry, consultants,
+communities, and interested individuals.
 
-User's without pre-existing knowledge of wildland fuels, spatial data, and R may need to dedicate more time and effort to learning how to apply the tools in this package by exploring the included documentation and linked resources. 
+User's without pre-existing knowledge of wildland fuels, spatial data,
+and R may need to dedicate more time and effort to learning how to apply
+the tools in this package by exploring the included documentation and
+linked resources.
 
 ### What the package does
 
-1.  Automates published methodologies
+#### Automate published methodologies
 
 This package was developed to automate the methods published in a series
 of scientific publications. Replicating the methods from a scientific
 publication can be challenging; it can require a significant amount of
 time and experience which can be a barrier to access.
 
-2.  Provides an accessible platform
+#### Provide an accessible platform
 
 This package and the code within it will always be free. Financial costs
 can also be a significant barrier to conducting wildfire risk
@@ -75,17 +87,22 @@ novice R users should have enough resources to conduct their own
 wildfire exposure assessments by referencing the documentation in this
 package and the plethora of free R tutorials available online.
 
-3.  Visualizes outputs
+#### Visualize outputs
 
-The outputs from the analysis functions can be quickly visualized in R with a selection of visualization functions that return plots, maps, or tables which can be exported as images or .csv files.
+The outputs from the analysis functions can be quickly visualized in R
+with a selection of visualization functions that return plots, maps, or
+tables which can be exported as images or .csv files.
 
-4.  Allows customization
+#### Allow customization
 
-Functions allow for custom parameterization to alter the analysis if desired. This allows for easy adaptation of the methods to suit the unique requirements for different use cases, geographic areas, and scales of analysis.  
+Functions allow for custom parameterization to alter the analysis if
+desired. This allows for easy adaptation of the methods to suit the
+unique requirements for different use cases, geographic areas, and
+scales of analysis.
 
 ### What the package doesn't
 
-1.  Let you skip the hard part (understanding the theory)
+#### Let you skip the hard part (understanding the theory)
 
[…]
