+++ pkgdown.yml
@@ -5,7 +5,8 @@ articles:
   fireexposuR: fireexposuR.html
   prep-input-data: prep-input-data.html
   prometheus: prometheus.html
-last_built: 2025-01-31T03:56Z
+  resources: resources.html
+last_built: 2025-01-31T20:41Z
 urls:
   reference: https://heyairf.github.io/fireexposuR/reference
   article: https://heyairf.github.io/fireexposuR/articles
