+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   prep-input-data: prep-input-data.html
   prometheus: prometheus.html
   resources: resources.html
-last_built: 2025-01-31T20:41Z
+last_built: 2025-01-31T23:18Z
 urls:
   reference: https://heyairf.github.io/fireexposuR/reference
   article: https://heyairf.github.io/fireexposuR/articles
