### 2025-02-02 #659 @maurolepore
Dear @heyairf this is to mark the start of my EiC rotation and record what I see from the perspective of that role.

@heyairf I see you've done a lot of work already. I also see you intend to work more and that's great. However I suggest we first focus on just-enough to satisfy the reviewers' requests. That way we can use the context they have in their heads before they loose it. Anything beyond that is important but not urgent and I encourage you to do it after approval.

In other words, when do you think you would be able to do the minimum required to satisfy the key issues the reviewers requested?

No problem if you think it'll take quite some time. We do understand and offer a holding label if you still choose to wait.

---

https://devguide.ropensci.org/softwarereview_policies.html#policiesreviewprocess

The author can choose to have their submission put on hold (editor applies the holding label). The holding status will be revisited every 3 months, and after one year the issue will be closed.
