### 2025-02-04 #659 @heyairf
Hi @maurolepore ,

I was just coming back to let everyone know that I am satisfied with the current state of the package, and that I think all reviewer comments have been addressed.

The only thing I am working on still is a final vignette that does not document the package or the functions within, but will offer advice for preparing input data.
