### 2025-02-08 #659 @ronnyhdez
## Reviewer Response

#### Final approval (post-review)

Estimated hours spent reviewing: 1

Note: cool hex sticker!
