### 2025-02-10 #659 @ropensci-review-bot
Approved! Thanks @heyairf for submitting and @ronnyhdez, @huizezhang-sherry for your reviews! :grin:
