### 2025-02-10 #659 @maurolepore
@heyairf , @ronnyhdez, and @huizezhang-sherry,

On a less automatic note, thanks so much for kindly contributing your work and experience. I look forward to seeing you around in the rOpenSci workspace on Slack. If for any reason you lack access please let me know.

All the best,
Mauro
