### 2025-02-18 #659 @heyairf
hi @maurolepore !! I am so excited. Thanks for all your help through the process (and @adamhsparks too! )
Thank you also for the to do list... I have joined the organization as it had the time limit, but haven't seen to any of the other tasks yet. If I have questions about them, is this the best place to ask? I will be prioritizing them this week!

Air
