### 2025-02-24 #659 @heyairf
Hi @maurolepore  ...

I think I've handled most of the to-do list but I have some questions about the pkgdown stuff....

I deleted the .yaml workflow that was added when I used [use_github_action("pkgdown")](https://usethis.r-lib.org/reference/use_github_action.html) which I THINK is what is the auto deployment for my pkgdown site that I was requested to remove with [this commit](https://github.com/ropensci/fireexposuR/commit/22114c0269561a5fe6df26e85a5ab2f8a75e96cb) . Was this what that bullet point meant? Or did I misinterpret it?

Now the few changes I've made to the vignettes with [this commit](https://github.com/ropensci/fireexposuR/commit/7e0c2b4f3d08a97f43a6a607fc39a32d0693435f) are not reflected in the site. Is there something else I have to change to make the page rebuild now that it has been transferred? I've read [this page](https://devguide.ropensci.org/pkg_building.html#docsropensci) but I am still confused about this!

Thank you!
Air
