### 2025-02-24 #659 @maurolepore
Thanks @heyairf  for your progress!

Yep, looks like the website is up and running! Congratulations! I guess the bots took a while to build the site but glad the site now reflects what you expect!

What I see here looks great!: https://docs.ropensci.org/fireexposuR/
