### 2025-02-24 #659 @heyairf
Amazing!

Next (and hopefully last) question.. the package is now failing to build for MacOS here: https://github.com/r-universe/ropensci/actions/runs/13508978097

when i run the checks or build locally on my Mac there are no errors or warnings...
How and where do I begin to figure out how to fix this?
