### 2025-02-25 #659 @maelle
Maybe some spatial deps aren't installed correctly on macOS on R-universe? @jeroen
