### 2025-02-25 #659 @maelle
Is it possible to run a "situation report" function on a machine to assess whether the spatial libraries have been installed correctly?
