### 2025-02-25 #659 @maurolepore
@heyairf

Thanks for following up.

At the [link](https://mac.r-project.org/macbuilder/submit.html) jeroen shared I see

> 1. You can upload a source R package that has you listed as the maintainer.
> 2. It must be prepared using R CMD build.

RE 1. Please try the instructions below yourself. If you can't let me know (I guess I can temporarily change the maintainer to myself and do it on your behalf).

RE 2. From the Terminal change directory to your package (`cd fireexposuR`) then build the package (`R CMD .`). This should create a file that you can then upload to https://mac.r-project.org/macbuilder/submit.html and when the process completes please share the link to the results. I'm not sure how long they'll persist so maybe just in case also paste the text you see in the "Build log" and "Check log" tabs. Here's an example from a package of my own that I uploaded to test the system:

![Image](https://github.com/user-attachments/assets/ce9333d7-20b5-4884-a22a-9848b0c9060d)

Build log

```
Resetting rust
gcc -o watchdog src/watchdog.c -O3
gcc -o sched src/sched.c -O3
gcc -Wall -o should_build src/should_build.c -O3
mkdir CRAN.meta
mkdir -p CRAN/src/contrib
*** use 'sync.bioc.in' to run CRAN+BioC sync
/Library/Developer/CommandLineTools/usr/bin/make  -C CRAN -f ../Makefile.CRAN cache
ln -s ../CRAN.meta meta
mkdir -p meta/src/contrib
tar fxzO src/contrib/dverse_0.2.0.9001.tar.gz `echo src/contrib/dverse_0.2.0.9001.tar.gz | sed -e 's:.*/::' -e 's:_.*::'`/DESCRIPTION > meta/src/contrib/dverse_0.2.0.9001.DESCRIPTION
../rmdup 'meta/src/contrib/*.DESCRIPTION'
RBUILDS: /Volumes/PkgBuild/work/1740478895-ab162a67b0575624/packages
base: /Volumes/PkgBuild/work/1740478895-ab162a67b0575624/packages/CRAN
sources: /Volumes/PkgBuild/work/1740478895-ab162a67b0575624/packages/CRAN/src/contrib
meta: /Volumes/PkgBuild/work/1740478895-ab162a67b0575624/packages/CRAN/meta/src/contrib
RBIN: R
CLANGPATH:
 [X11] Using session X at :0
R version 4.4.2
 SKIP_CHK: no
 CHK_ONLY: no
 UPDATE  : no
OUTPUT:
  Rlib: /Volumes/PkgBuild/work/1740478895-ab162a67b0575624/packages/big-sur-arm64/Rlib/4.4
  bin : /Volumes/PkgBuild/work/1740478895-ab162a67b0575624/packages/big-sur-arm64/bin/4.4
  res : /Volumes/PkgBuild/work/1740478895-ab162a67b0575624/packages/big-sur-arm64/results/4.4
RLIBS=/Volumes/PkgBuild/work/1740478895-ab162a67b0575624/packages/big-sur-arm64/Rlib/4.4:/Volumes/PkgBuild/work/1740478895-ab162a67b0575624/packages/big-sur-arm64/CRANlib/4.4
Machine : Mac mini, Apple M1
System  : macOS 13.3.1 (22E261)
Profile : r-release-macosx-arm64
SDK     : macOS 11.3 (C compiler: clang-1403.0.22.14.1)

Building package dependency tree..
(from /Volumes/PkgBuild/work/1740478895-ab162a67b0575624/packages/CRAN/src/contrib)
./dtree --desc /Volumes/PkgBuild/work/1740478895-ab162a67b0575624/packages/CRAN/meta/src/contrib /Volumes/PkgBuild/work/1740478895-ab162a67b0575624/packages/CRAN/src/contrib > /Volumes/PkgBuild/work/1740478895-ab162a67b0575624/packages/CRAN/dep.list
NOTE: there is no source for package DT [DT], installed CRAN version will be used.
NOTE: there is no source for package knitr [knitr], installed CRAN version will be used.
NOTE: there is no source for package testthat [testthat], installed CRAN version will be used.
NOTE: there is no source for package withr [withr], installed CRAN version will be used.
NOTE: there is no source for package cli [cli], installed CRAN version will be used.
NOTE: there is no source for package curl [curl], installed CRAN version will be used.
NOTE: there is no source for package dplyr [dplyr], installed CRAN version will be used.
NOTE: there is no source for package glue [glue], installed CRAN version will be used.
NOTE: there is no source for package rlang [rlang], installed CRAN version will be used.
NOTE: there is no source for package tibble [tibble], installed CRAN version will be used.
Checking all packages..
1740478913:20250225:232153:dverse:
  /Volumes/PkgBuild/work/1740478895-ab162a67b0575624/packages/CRAN/src/contrib/dverse_0.2.0.9001.tar.gz
 - looking for /Volumes/PkgBuild/work/1740478895-ab162a67b0575624/packages/CRAN/meta/src/contrib/dverse_0.2.0.9001.DESCRIPTION
 - using cached DESCRIPTION file for detection
  - unpacking
  - checking dverse
  - installing from sources
* installing *binary* package ‘dverse’ ...
* DONE (dverse)
```

Check log

```
* using log directory ‘/Volumes/PkgBuild/work/1740478895-ab162a67b0575624/packages/big-sur-arm64/results/4.4/dverse.Rcheck’
* using R version 4.4.2 (2024-10-31)
* using platform: aarch64-apple-darwin20
* R was compiled by
    Apple clang version 14.0.0 (clang-1400.0.29.202)
    GNU Fortran (GCC) 12.2.0
* running under: macOS Ventura 13.3.1
* using session charset: UTF-8
* checking for file ‘dverse/DESCRIPTION’ ... OK
* this is package ‘dverse’ version ‘0.2.0.9001’
* package encoding: UTF-8
* checking package namespace information ... OK
* checking package dependencies ... OK
* checking if this is a source package ... OK
* checking if there is a namespace ... OK
* checking for executable files ... OK
* checking for hidden files and directories ... OK
* checking for portable file names ... OK
* checking for sufficient/correct file permissions ... OK
* checking whether package ‘dverse’ can be installed ... [1s/1s] OK
* checking installed package size ... OK
* checking package directory ... OK
* checking top-level files ... OK
* checking for left-over files ... OK
* checking index information ... OK
* checking package subdirectories ... OK
* checking code files for non-ASCII characters ... OK
* checking R files for syntax errors ... OK
* checking whether the package can be loaded ... [0s/0s] OK
* checking whether the package can be loaded with stated dependencies ... [0s/0s] OK
* checking whether the package can be unloaded cleanly ... [0s/0s] OK
* checking whether the namespace can be loaded with stated dependencies ... [0s/0s] OK
* checking whether the namespace can be unloaded cleanly ... [0s/0s] OK
* checking loading without being on the library search path ... [0s/0s] OK
* checking dependencies in R code ... OK
* checking S3 generic/method consistency ... OK
* checking replacement functions ... OK
* checking foreign function calls ... OK
* checking R code for possible problems ... [1s/1s] OK
* checking Rd files ... [0s/0s] OK
* checking Rd metadata ... OK
* checking Rd cross-references ... OK
* checking for missing documentation entries ... OK
* checking for code/documentation mismatches ... OK
* checking Rd \usage sections ... OK
* checking Rd contents ... OK
* checking for unstated dependencies in examples ... OK
* checking examples ... [34s/44s] OK
* checking for unstated dependencies in ‘tests’ ... OK
* checking tests ... [63s/65s] OK
  Running ‘testthat.R’ [63s/65s]
* checking PDF version of manual ... [2s/2s] OK
* DONE
Status: OK
* using check arguments '--no-clean-on-error '

* elapsed time (check, wall clock): 2:03

```

BTW in case you're wondering, you can create the drop-down thingy above with

````

```
paste your text here
```

````
