### 2025-02-25 #659 @heyairf
Thanks everyone for your help ! Here are the results following those instructions...

Build Log:

```
Resetting rust
gcc -o watchdog src/watchdog.c -O3
gcc -o sched src/sched.c -O3
gcc -Wall -o should_build src/should_build.c -O3
mkdir CRAN.meta
mkdir -p CRAN/src/contrib
*** use 'sync.bioc.in' to run CRAN+BioC sync
/Library/Developer/CommandLineTools/usr/bin/make  -C CRAN -f ../Makefile.CRAN cache
ln -s ../CRAN.meta meta
mkdir -p meta/src/contrib
tar fxzO src/contrib/fireexposuR_1.1.0.tar.gz `echo src/contrib/fireexposuR_1.1.0.tar.gz | sed -e 's:.*/::' -e 's:_.*::'`/DESCRIPTION > meta/src/contrib/fireexposuR_1.1.0.DESCRIPTION
../rmdup 'meta/src/contrib/*.DESCRIPTION'
RBUILDS: /Volumes/PkgBuild/work/1740512281-be2e6ab4a40f63f7/packages
base: /Volumes/PkgBuild/work/1740512281-be2e6ab4a40f63f7/packages/CRAN
sources: /Volumes/PkgBuild/work/1740512281-be2e6ab4a40f63f7/packages/CRAN/src/contrib
meta: /Volumes/PkgBuild/work/1740512281-be2e6ab4a40f63f7/packages/CRAN/meta/src/contrib
RBIN: R
CLANGPATH:
 [X11] Using session X at :0
R version 4.4.2
 SKIP_CHK: no
 CHK_ONLY: no
 UPDATE  : no
OUTPUT:
  Rlib: /Volumes/PkgBuild/work/1740512281-be2e6ab4a40f63f7/packages/big-sur-arm64/Rlib/4.4
  bin : /Volumes/PkgBuild/work/1740512281-be2e6ab4a40f63f7/packages/big-sur-arm64/bin/4.4
  res : /Volumes/PkgBuild/work/1740512281-be2e6ab4a40f63f7/packages/big-sur-arm64/results/4.4
RLIBS=/Volumes/PkgBuild/work/1740512281-be2e6ab4a40f63f7/packages/big-sur-arm64/Rlib/4.4:/Volumes/PkgBuild/work/1740512281-be2e6ab4a40f63f7/packages/big-sur-arm64/CRANlib/4.4
Machine : Mac mini, Apple M1
System  : macOS 13.3.1 (22E261)
Profile : r-release-macosx-arm64
SDK     : macOS 11.3 (C compiler: clang-1403.0.22.14.1)

Building package dependency tree..
(from /Volumes/PkgBuild/work/1740512281-be2e6ab4a40f63f7/packages/CRAN/src/contrib)
./dtree --desc /Volumes/PkgBuild/work/1740512281-be2e6ab4a40f63f7/packages/CRAN/meta/src/contrib /Volumes/PkgBuild/work/1740512281-be2e6ab4a40f63f7/packages/CRAN/src/contrib > /Volumes/PkgBuild/work/1740512281-be2e6ab4a40f63f7/packages/CRAN/dep.list
NOTE: there is no source for package knitr [knitr], installed CRAN version will be used.
NOTE: there is no source for package rmarkdown [rmarkdown], installed CRAN version will be used.
NOTE: there is no source for package testthat [testthat], installed CRAN version will be used.
NOTE: there is no source for package dplyr [dplyr], installed CRAN version will be used.
NOTE: there is no source for package geosphere [geosphere], installed CRAN version will be used.
NOTE: there is no source for package ggplot2 [ggplot2], installed CRAN version will be used.
NOTE: there is no source for package ggspatial [ggspatial], installed CRAN version will be used.
NOTE: there is no source for package magrittr [magrittr], installed CRAN version will be used.
NOTE: there is no source for package maptiles [maptiles], installed CRAN version will be used.
NOTE: there is no source for package MultiscaleDTM [MultiscaleDTM], installed CRAN version will be used.
NOTE: there is no source for package rlang [rlang], installed CRAN version will be used.
NOTE: there is no source for package terra [terra], installed CRAN version will be used.
NOTE: there is no source for package tidyr [tidyr], installed CRAN version will be used.
NOTE: there is no source for package tidyselect [tidyselect], installed CRAN version will be used.
NOTE: there is no source for package tidyterra [tidyterra], installed CRAN version will be used.
Checking all packages..
1740512297:20250226:083817:fireexposuR:
  /Volumes/PkgBuild/work/1740512281-be2e6ab4a40f63f7/packages/CRAN/src/contrib/fireexposuR_1.1.0.tar.gz
 - looking for /Volumes/PkgBuild/work/1740512281-be2e6ab4a40f63f7/packages/CRAN/meta/src/contrib/fireexposuR_1.1.0.DESCRIPTION
 - using cached DESCRIPTION file for detection
  - unpacking
  - checking fireexposuR
  - installing from sources
* installing *binary* package ‘fireexposuR’ ...
* DONE (fireexposuR)
```

Check Log:

```
* using log directory ‘/Volumes/PkgBuild/work/1740512281-be2e6ab4a40f63f7/packages/big-sur-arm64/results/4.4/fireexposuR.Rcheck’
* using R version 4.4.2 (2024-10-31)
* using platform: aarch64-apple-darwin20
* R was compiled by
    Apple clang version 14.0.0 (clang-1400.0.29.202)
    GNU Fortran (GCC) 12.2.0
* running under: macOS Ventura 13.3.1
* using session charset: UTF-8
* checking for file ‘fireexposuR/DESCRIPTION’ ... OK
* checking extension type ... Package
* this is package ‘fireexposuR’ version ‘1.1.0’
* package encoding: UTF-8
* checking package namespace information ... OK
* checking package dependencies ... OK
* checking if this is a source package ... OK
* checking if there is a namespace ... OK
* checking for executable files ... OK
* checking for hidden files and directories ... OK
* checking for portable file names ... OK
* checking for sufficient/correct file permissions ... OK
* checking whether package ‘fireexposuR’ can be installed ... [1s/1s] OK
* checking installed package size ... OK
* checking package directory ... OK
* checking ‘build’ directory ... OK
* checking top-level files ... OK
* checking for left-over files ... OK
* checking index information ... OK
* checking package subdirectories ... OK
* checking code files for non-ASCII characters ... OK
* checking R files for syntax errors ... OK
* checking whether the package can be loaded ... [0s/0s] OK
* checking whether the package can be loaded with stated dependencies ... [0s/0s] OK
* checking whether the package can be unloaded cleanly ... [0s/0s] OK
* checking whether the namespace can be loaded with stated dependencies ... [0s/0s] OK
* checking whether the namespace can be unloaded cleanly ... [0s/0s] OK
* checking loading without being on the library search path ... [0s/0s] OK
* checking dependencies in R code ... OK
* checking S3 generic/method consistency ... OK
* checking replacement functions ... OK
* checking foreign function calls ... OK
* checking R code for possible problems ... [1s/1s] OK
* checking Rd files ... [0s/0s] OK
* checking Rd metadata ... OK
* checking Rd cross-references ... OK
* checking for missing documentation entries ... OK
* checking for code/documentation mismatches ... OK
* checking Rd \usage sections ... OK
* checking Rd contents ... OK
* checking for unstated dependencies in examples ... OK
* checking installed files from ‘inst/doc’ ... OK
* checking files in ‘vignettes’ ... OK
* checking examples ... [8s/12s] OK
* checking for unstated dependencies in ‘tests’ ... OK
* checking tests ... [12s/16s] OK
  Running ‘testthat.R’ [12s/16s]
* checking for unstated dependencies in vignettes ... OK
* checking package vignettes ... OK
* checking re-building of vignette outputs ... [5s/6s] OK
* checking PDF version of manual ... [2s/2s] OK
* DONE
Status: OK
* using check arguments '--no-clean-on-error '

* elapsed time (check, wall clock): 0:50
```

I built from both my mac laptop and from my PC desktop and got the same results.
