### 2025-02-25 #659 @jeroen
I don't have time to debug it now, let's ignore it. Perhaps there is some issue with the GDAL version on the r-universe server. FWIW on r-universe your package builds fine and is being published as normal, even if a test fails.
